### 2023-12-23 935062c
fix: [44] fixed wrong type hint; isort; black formatting; trim trailing whitespaces
