### 2023-12-23 ae352b3
feat: [44] added ruff linting checks; added requirements for development
