### 2023-12-23 67faaf3
feat: [44] added finite difference computation for Whittaker smoothers
