### 2023-12-23 d13c5ce
feat: [44] added more banded linear algebra functions with `scipy`s LAPACK wrappers for Whittaker smoothers
