### 2023-12-23 0041382
feat: [44] added centralized Whittaker smoother class
