+++ chemotools/utils/whittaker_base.py
@@ -1,0 +1,626 @@
+import numpy as np
+from scipy.linalg import cho_solve_banded, cholesky_banded
+from scipy.optimize import minimize_scalar
+
+from chemotools.utils.banded_linalg import (
+    conv_symm_sparse_banded_sposdef_to_posdef,
+    conv_to_lu_banded_storage,
+    lu_banded,
+    lu_solve_banded,
+    slodget_cho_banded,
+)
+from chemotools.utils.finite_differences import forward_finite_diff_conv_matrix
+from chemotools.utils.models import _PENTAPY_AVAILABLE, BandedSolveDecompositions
+
+if _PENTAPY_AVAILABLE:
+    import pentapy as pp
+# else nothing
+
+_LUDecomposition = tuple[np.ndarray, np.ndarray, tuple[int, int]]
+_CholeskyDecomposition = tuple[np.ndarray, bool]
+_PentapyDecomposition = None
+_Decomposition = _PentapyDecomposition | _CholeskyDecomposition | _LUDecomposition
+
+LN_OF_TWO_PI = np.log(2.0 * np.pi)
+
+
+class WhittakerLikeSolver:
+    """This class can be used to solve linear systems of equations that involve banded
+    matrices as they occur in applications like the Whittaker-Henderson-smoothing or
+    derived methods like Asymmetric Least Squares (ALS) baseline correction.
+    It support weights and tries to use the most efficient method available.
+    Besides, it also offers the possibility to fit the roughness penalty itself.
+
+    Parameters
+    ----------
+    lam : int or float or None, default=1e2
+        The lambda parameter to use for the Whittaker smooth.
+        If ``None``, the transformer will fit the smoothness parameter itself by
+        maximising the marginal likelihood, which can be computationally expensive, but
+        more accurate than using (Generalized) Cross-Validation (see Notes).
+
+    differences : int, default=1
+        The number of differences to use for the Whittaker smooth. If the aim is to
+        obtain a smooth estimate of the `m`-th order derivative, this should be set to
+        at least ``m + 2``.
+
+    polynomial_order : int, default=1
+        Same as ``differences``, but for ``AirPls``.
+
+    rcond : float, default=1e-15
+        The relative condition number which is used to keep all matrices involved
+        positive definite. This is only used if ``lam`` is ``None``.
+        It works in the same way as the ``rcond`` parameter of SciPy's ``linalg.pinvh``.
+
+    Notes
+    -----
+    If ``lam`` is ``None``, the pentapy solver cannot be used even if available.
+    Besides, the computational load increases since - especially for large series and
+    high differences - the pseudo-determinant of the penalty matrix ``P = D^T @ D``
+    needs to be computed, which is computationally expensive and also subject to
+    numerical inaccuracies. The latter cause some eigenvalues to be numerically
+    negative, even though their true value is positive and of order ``<< 1e-16``, which
+    makes their accurate computation numerically impossible with double precision.
+    Therefore, the eigenvalues are lifted by adding a small value to the diagonal of
+    the penalty matrix before computing the pseudo-determinant to make the smallest
+    eigenvalue numerically positive when compared to the largest eigenvalue. From a
+    smoothing point of view, this turns the Whittaker-Smoothing with derivative penalty
+    into a blend of Whittaker Smoothing and Tikhonov Regularisation.
+    So, in contrast to ``P = D^T @ D``, the penalty matrix ``P = D^T @ D + c * I`` is
+    used where ``c`` is a very small numerical value, so in first approximation, the
+    combined smoother is still mostly a Whittaker smoother.
+
+    """
+
+    __log_lam_bounds = (
+        -34.5,  # 1e-15
+        115.13,  # 1e50
+    )
+
+    def __init__(
+        self,
[…]
