### 2023-12-23 c3dff48
feat: [44] unified whittaker-smoother-like methods with performance enhancement; extended tests; fixed type hints; black formatted; started lint fixes
