+++ chemotools/baseline/_ar_pls.py
@@ -1,18 +1,17 @@
 import logging
-import numpy as np
-import scipy.sparse as sp
-from scipy.sparse import spdiags, csc_matrix
-from scipy.sparse.linalg import splu
+from numbers import Integral
 
-from sklearn.base import BaseEstimator, TransformerMixin, OneToOneFeatureMixin
-from sklearn.utils.validation import check_is_fitted
+import numpy as np
+from sklearn.base import BaseEstimator, OneToOneFeatureMixin, TransformerMixin
+from sklearn.utils.validation import check_is_fitted, check_scalar
 
 from chemotools.utils.check_inputs import check_input
+from chemotools.utils.whittaker_base import WhittakerLikeSolver
 
 logger = logging.getLogger(__name__)
 
 
-class ArPls(OneToOneFeatureMixin, BaseEstimator, TransformerMixin):
+class ArPls(OneToOneFeatureMixin, BaseEstimator, TransformerMixin, WhittakerLikeSolver):
     """
     This class implements the Assymmetrically Reweighted Penalized Least Squares (ArPls) is a baseline
     correction method for spectroscopy data. It uses an iterative process
@@ -20,7 +19,7 @@ class ArPls(OneToOneFeatureMixin, BaseEstimator, TransformerMixin):
 
     Parameters
     ----------
-    lam : float, optional (default=1e4)
+    lam : float or int, optional (default=1e4)
         The penalty parameter for the difference matrix in the objective function.
 
     ratio : float, optional (default=0.01)
@@ -28,6 +27,11 @@ class ArPls(OneToOneFeatureMixin, BaseEstimator, TransformerMixin):
 
     nr_iterations : int, optional (default=100)
         The maximum number of iterations for the weight updating scheme.
+
+    rcond : float, default=1e-15
+        The relative condition number which is used to keep all matrices involved
+        positive definite. This is not actively used at the moment.
+        It works in the same way as the ``rcond`` parameter of SciPy's ``linalg.pinvh``.
 
 
     Methods
@@ -46,20 +50,24 @@ class ArPls(OneToOneFeatureMixin, BaseEstimator, TransformerMixin):
 
     References
     ----------
-    - Sung-June Baek, Aaron Park, Young-Jin Ahn, Jaebum Choo 
-    Baseline correction using asymmetrically reweighted penalized 
+    - Sung-June Baek, Aaron Park, Young-Jin Ahn, Jaebum Choo
+    Baseline correction using asymmetrically reweighted penalized
     least squares smoothing
     """
 
     def __init__(
         self,
-        lam: float = 1e4,
+        lam: float | int = 1e4,
+        differences: int = 2,
         ratio: float = 0.01,
         nr_iterations: int = 100,
+        rcond: float = 1e-15,
     ):
-        self.lam = lam
-        self.ratio = ratio
-        self.nr_iterations = nr_iterations
+        self.lam: float | int = lam
+        self.differences: int = differences
+        self.ratio: float = ratio
+        self.nr_iterations: int = nr_iterations
+        self.rcond: float = rcond
 
     def fit(self, X: np.ndarray, y=None) -> "ArPls":
         """Fit the estimator to the data.
@@ -78,8 +86,28 @@ class ArPls(OneToOneFeatureMixin, BaseEstimator, TransformerMixin):
             Returns the instance itself.
         """
 
+        # the constructor parameters are checked
+        check_scalar(
+            x=self.ratio,
+            name="ratio",
+            target_type=float,
+            min_val=1e-15,
+            max_val=1.0 - 1e-15,
+        )
+        check_scalar(
+            x=self.nr_iterations, name="nr_iterations", target_type=Integral, min_val=1
+        )
+
[…]
