+++ chemotools/smooth/_whittaker_smooth.py
@@ -1,46 +1,81 @@
-import numpy as np
-from scipy.sparse import csc_matrix, eye, diags
-from scipy.sparse.linalg import spsolve
-from sklearn.base import BaseEstimator, TransformerMixin, OneToOneFeatureMixin
+"""
+This module contains the ``WhittakerSmooth`` transformer, which performs smoothing on
+data according to the Whittaker-Henderson formulation of Penalized Least Squares.
+
+References
+----------
+It's based on the algorithms described in [1]_ and [2]_ where an implementational
+adaption of [2]_ was required to make it numerically stable ([3]_).
+
+.. [1] Z.-M. Zhang, S. Chen, and Y.-Z. Liang, "Baseline correction using adaptive
+   iteratively reweighted penalized least squares", Analyst 135 (5), 1138-1146 (2010).
+.. [2] G. Biessy, "Revisiting Whittaker-Henderson smoothing", arXiv:2306.06932 (2023).
+.. [3] https://math.stackexchange.com/q/4819039/1261538
+
+"""
+
+from numpy import ndarray
+from sklearn.base import BaseEstimator, OneToOneFeatureMixin, TransformerMixin
 from sklearn.utils.validation import check_is_fitted
 
-from chemotools.utils.check_inputs import check_input
-
-# This code is adapted from the following source:
-# Z.-M. Zhang, S. Chen, and Y.-Z. Liang, 
-# Baseline correction using adaptive iteratively reweighted penalized least squares. 
-# Analyst 135 (5), 1138-1146 (2010).
+from chemotools.utils.check_inputs import check_input, check_weights
+from chemotools.utils.whittaker_base import WhittakerLikeSolver
 
 
-class WhittakerSmooth(OneToOneFeatureMixin, BaseEstimator, TransformerMixin):
+class WhittakerSmooth(
+    OneToOneFeatureMixin, BaseEstimator, TransformerMixin, WhittakerLikeSolver
+):
     """
-    A transformer that calculates the Whittaker smooth of the input data.
+    A transformer that performs smoothing on data according to the Whittaker-Henderson
+    formulation of Penalized Least Squares.
 
     Parameters
     ----------
-    lam : float, optional
-        The lambda parameter to use for the Whittaker smooth. Default is 1e2.
+    lam : float or int, default=1e2
+        The lambda parameter to use for the Whittaker smooth.
 
-    differences : int, optional
-        The number of differences to use for the Whittaker smooth. Default is 1.
+    differences : int, default=1
+        The number of differences to use for the Whittaker smooth. If the aim is to
+        obtain a smooth estimate of the `m`-th order derivative, this should be set to
+        at least ``m + 2``.
+
+    rcond : float, default=1e-15
+        The relative condition number which is used to keep all matrices involved
+        positive definite. This is not actively used at the moment.
+        It works in the same way as the ``rcond`` parameter of SciPy's ``linalg.pinvh``.
+
+    Attributes
+    ----------
+    n_features_in_ : int
+        The number of features in the input data.
+
+    _is_fitted : bool
+        Whether the transformer has been fitted to data.
 
     Methods
     -------
     fit(X, y=None)
         Fit the transformer to the input data.
 
-    transform(X, y=0, copy=True)
-        Transform the input data by calculating the Whittaker smooth.
+    transform(X, y=None, sample_weight=None)
+        Transform the input data by calculating the (weighted) Whittaker smooth.
+
+    fit_transform(X, y=None, sample_weight=None)
+        Fit the transformer to the input data and transform it.
+
     """
+
     def __init__(
         self,
-        lam: float = 1e2,
+        lam: int | float = 1e2,
         differences: int = 1,
+        rcond: float = 1e-15,
     ):
         self.lam = lam
[…]

[…]
