### 2023-12-23 995a964
style: [44] made docstrings of whittaker smoothers black-compatible; added references
