+++ chemotools/baseline/_ar_pls.py
@@ -1,3 +1,23 @@
+"""
+This module contains the ``ArPLS`` transformer, which performs baseline correction on
+data according to the Whittaker-Henderson formulation of Penalized Least Squares which
+was modified by the introduction of weights that are updated iteratively to improve the
+baseline identification. It simultaneously estimates the baseline as well as the
+baseline noise.
+
+References
+----------
+It's based on the algorithms described in [1]_ and [2]_ where an implementational
+adaption of [2]_ was required to make it numerically stable ([3]_).
+
+.. [1] S.-J. Baek, A. Park, Y.-J. Ahn, J. Choo, "Baseline correction using
+   asymmetrically reweighted penalized least squares smoothing", Analyst, 140, 250–257
+   (2015)
+.. [2] G. Biessy, "Revisiting Whittaker-Henderson smoothing", arXiv:2306.06932 (2023)
+.. [3] https://math.stackexchange.com/q/4819039/1261538
+
+"""
+
 import logging
 from numbers import Integral
 
@@ -13,17 +33,21 @@ logger = logging.getLogger(__name__)
 
 class ArPls(OneToOneFeatureMixin, BaseEstimator, TransformerMixin, WhittakerLikeSolver):
     """
-    This class implements the Assymmetrically Reweighted Penalized Least Squares (ArPls) is a baseline
-    correction method for spectroscopy data. It uses an iterative process
-    to estimate and remove the baseline from the spectra.
+    This class implements the Asymmetrically Reweighted Penalized Least Squares a.k.a
+    ArPLS which is a baseline correction method for spectroscopy data. It uses an
+    iterative process that simultaneously estimates the baseline as well as the baseline
+    noise.
 
     Parameters
     ----------
-    lam : float or int, optional (default=1e4)
-        The penalty parameter for the difference matrix in the objective function.
+    lam : float or int, default=1e4
+        The lambda parameter that controls the smoothness of the baseline. Higher values
+        will result in a smoother baseline.
 
-    ratio : float, optional (default=0.01)
-        The convergence threshold for the weight updating scheme.
+    ratio : float, default=0.01
+        The convergence threshold for the weight updating scheme. Lower values will
+        result in a more accurate baseline at the cost of computation time and even
+        convergence.
 
     nr_iterations : int, optional (default=100)
         The maximum number of iterations for the weight updating scheme.
@@ -33,7 +57,6 @@ class ArPls(OneToOneFeatureMixin, BaseEstimator, TransformerMixin, WhittakerLike
         positive definite. This is not actively used at the moment.
         It works in the same way as the ``rcond`` parameter of SciPy's ``linalg.pinvh``.
 
-
     Methods
     -------
     fit(X, y=None)
@@ -42,17 +65,18 @@ class ArPls(OneToOneFeatureMixin, BaseEstimator, TransformerMixin, WhittakerLike
     transform(X, y=None)
         Transform the data by removing the baseline.
 
-    _calculate_diff(N)
-        Calculate the difference matrix for a given size.
-
     _calculate_ar_pls(x)
         Calculate the baseline for a given spectrum.
 
     References
     ----------
-    - Sung-June Baek, Aaron Park, Young-Jin Ahn, Jaebum Choo
-    Baseline correction using asymmetrically reweighted penalized
-    least squares smoothing
+    .. [1] S.-J. Baek, A. Park, Y.-J. Ahn, J. Choo, "Baseline correction using
+       asymmetrically reweighted penalized least squares smoothing", Analyst, 140,
+       250–257 (2015)
[…]
