### 2023-12-25 0437cc7
wip: [44] figured out pentapy problems and saved working status for figuring out 100% fix
