+++ chemotools/utils/whittaker_base.py
@@ -31,26 +31,27 @@ class WhittakerLikeSolver:
     It support weights and tries to use the most efficient method available.
     Besides, it also offers the possibility to fit the roughness penalty itself.
 
-    Parameters
+    Attributes
     ----------
-    lam : int or float or None, default=1e2
+    _lam : int or float or None, default=1e2
         The lambda parameter to use for the Whittaker smooth.
         If ``None``, the transformer will fit the smoothness parameter itself by
         maximising the marginal likelihood, which can be computationally expensive, but
         more accurate than using (Generalized) Cross-Validation (see Notes).
 
-    differences : int, default=1
+    _differences : int, default=1
         The number of differences to use for the Whittaker smooth. If the aim is to
         obtain a smooth estimate of the `m`-th order derivative, this should be set to
         at least ``m + 2``.
 
-    polynomial_order : int, default=1
-        Same as ``differences``, but for ``AirPls``.
-
-    rcond : float, default=1e-15
+    _rcond : float, default=1e-15
         The relative condition number which is used to keep all matrices involved
         positive definite. This is only used if ``lam`` is ``None``.
         It works in the same way as the ``rcond`` parameter of SciPy's ``linalg.pinvh``.
+
+    _allow_pentapy : bool, default=True
+        Whether to enable the Pentapy solver if available. This is only used for
+        debugging and testing purposes.
 
     Notes
     -----
@@ -72,17 +73,16 @@ class WhittakerLikeSolver:
 
     """
 
-    __log_lam_bounds = (
+    __log_lam_bounds: tuple[float, float] = (
         -34.5,  # 1e-15
         115.13,  # 1e50
     )
+    __allow_pentapy: bool = True
 
     def __init__(
         self,
     ) -> None:
-        self._lam: int | float | None = float("nan")
-        self._differences: int = -1
-        self._rcond: float = float("nan")
+        pass
 
     def _setup_for_fit(
         self,
@@ -96,9 +96,9 @@ class WhittakerLikeSolver:
         """
 
         # the input arguments are stored
-        self._lam = lam
-        self._differences = differences
-        self._rcond = rcond
+        self._lam: int | float | None = lam
+        self._differences: int = differences
+        self._rcond: float = rcond
 
         # the banded storage for a LAPACK LU decomposition is computed for the squared
         # forward finite difference matrix D^T @ D which is the penalty matrix P
@@ -151,22 +151,48 @@ class WhittakerLikeSolver:
         # finally, Pentapy is enabled if available, the number of differences is 2,
         # and the lambda parameter is not fitted automatically
         self._pentapy_enabled: bool = (
-            _PENTAPY_AVAILABLE and self._differences == 2 and not self.auto_lam_
+            _PENTAPY_AVAILABLE
+            and self._differences == 2
+            and not self.auto_lam_
+            and self.__allow_pentapy
         )
 
     def _pentapy_solve(self, ab: np.ndarray, bw: np.ndarray) -> np.ndarray:
         """Solves the linear system of equations ``(W + lam * D^T @ D) @ x = W @ b``
         with the Pentapy package. This is written as the system ``A @ x = b`` where
         ``A = W + lam * D^T @ D`` and ``b = W @ b``.
+
+        Notes
+        -----
+        Pentapy does not (maybe yet) allow for 2D right-hand side matrices, so the
+        solution is computed for each column of ``bw`` separately.
+
         """
 
-        return pp.solve(
-            mat=ab,
-            rhs=bw,
-            is_flat=True,
-            index_row_wise=False,
-            solver=1,
-        )
+        # for 1-dimensional right-hand side vectors, the solution is computed directly
+        if bw.ndim == 1:
+            return pp.solve(
+                mat=ab,
+                rhs=bw,
[…]
