### 2023-12-27 f37ffe7
refactor: [44] changed paradigm in Whittaker smoothing to work with mixed Tikhonov regularisation; shotgun refactor of all respective modules; extended and parametrized tests; added more FIXMEs and TODOs; adapted docstrings
