+++ chemotools/utils/models.py
@@ -16,6 +16,4 @@ except ImportError:
 
 class BandedSolveDecompositions(str, Enum):
     CHOLESKY = "cholesky"
-    LU = "lu"
-    POLYFIT = "polyfit"
     PENTAPY = "pentapy"
