### 2023-12-27 b9fd2e0
refactor: [44] temporarily disabled long computation of squared finite difference matrix in favour of speed; added explanation of Whittaker smoothing
