### 2024-02-04 4106f45
update range_cut method docstrings
