+++ get-started/coffee_spectra_classifier.md
@@ -10,12 +10,15 @@ nav_order: 5
 
 ## What will you learn?
 
-- [The coffee dataset](#the-coffee-dataset-🌍)
-- [Importing the data](#importing-the-data)
-- [Explore, plot and color](#explore-plot-and-color)
-- [Exploring the data](#exploring-the-data-🤓)
-- [Preprocessing the spectra](#preprocessing-the-spectra-🌊)
-- [Modelling the data](#modelling-the-data)
+- [__Coffee Spectra Classifier__](#coffee-spectra-classifier)
+  - [What will you learn?](#what-will-you-learn)
+  - [__Unlocking the Secrets of Coffee: A Spectral Journey ☕__](#unlocking-the-secrets-of-coffee-a-spectral-journey-)
+    - [__The Coffee Dataset 🌍__](#the-coffee-dataset-)
+  - [__Importing the data__](#importing-the-data)
+  - [__Explore, plot and color__](#explore-plot-and-color)
+  - [__Preprocessing the spectra  🌊__](#preprocessing-the-spectra--)
+  - [__Modelling the data__](#modelling-the-data)
+  - [__Recap__](#recap)
 
 ## __Unlocking the Secrets of Coffee: A Spectral Journey ☕__
 
@@ -160,7 +163,7 @@ We will build the preprocessing steps in a pipeline using the ``` make_pipeline(
 
 - __[Derivative](https://paucablop.github.io/chemotools/docs/derivative.html#savitzky-golay-derivative)__ to remove both additive and multiplicative scattering effects.
 
-- __[Range cut](https://paucablop.github.io/chemotools/docs/variable_selection.html#range-cut)__ to select the most relevant wavenumbers.
+- __[Range cut](https://paucablop.github.io/chemotools/docs/feature_selection.html#range-cut)__ to select the most relevant wavenumbers.
 
 - __[Standardize](https://scikit-learn.org/stable/modules/generated/sklearn.preprocessing.StandardScaler.html)__ remove the mean from the dataset.
 
@@ -170,7 +173,7 @@ from sklearn.preprocessing import StandardScaler
 
 from chemotools.derivative import SavitzkyGolay
 from chemotools.scatter import StandardNormalVariate
-from chemotools.variable_selection import RangeCut
+from chemotools.feature_selection import RangeCut
 
 pipeline = make_pipeline(
     StandardNormalVariate(),
