### 2024-02-04 4dc5116
update regressor documentation
