### 2024-02-11 26ac414
extend polars support to provided datasets
