+++ tests/test_datasets.py
@@ -1,9 +1,15 @@
 import pandas as pd
+import polars as pl
+import pytest
 
-from chemotools.datasets import load_coffee, load_fermentation_test, load_fermentation_train
+from chemotools.datasets import (
+    load_coffee,
+    load_fermentation_test,
+    load_fermentation_train,
+)
 
 
-def test_load_coffee():
+def test_load_coffee_pandas():
     # Arrange
 
     # Act
@@ -16,7 +22,28 @@ def test_load_coffee():
     assert isinstance(coffee_labels, pd.DataFrame)
 
 
-def test_load_fermentation_test():
+def test_load_coffee_polars():
+    # Arrange
+
+    # Act
+    coffee_spectra, coffee_labels = load_coffee(set_output="polars")
+
+    # Assert
+    assert coffee_spectra.shape == (60, 1841)
+    assert coffee_labels.shape == (60, 1)
+    assert isinstance(coffee_spectra, pl.DataFrame)
+    assert isinstance(coffee_labels, pl.DataFrame)
+
+
+def test_load_coffee_exception():
+    # Arrange
+
+    # Act and Assert
+    with pytest.raises(ValueError):
+        coffee_spectra, coffee_labels = load_coffee(set_output="plars")
+
+
+def test_load_fermentation_test_pandas():
     # Arrange
 
     # Act
@@ -28,7 +55,29 @@ def test_load_fermentation_test():
     assert isinstance(test_spectra, pd.DataFrame)
     assert isinstance(test_hplc, pd.DataFrame)
 
-def test_load_fermentation_train():
+
+def test_load_fermentation_test_polars():
+    # Arrange
+
+    # Act
+    test_spectra, test_hplc = load_fermentation_test(set_output="polars")
+
+    # Assert
+    assert test_spectra.shape == (1629, 1047)
+    assert test_hplc.shape == (34, 6)
+    assert isinstance(test_spectra, pl.DataFrame)
+    assert isinstance(test_hplc, pl.DataFrame)
+
+
+def test_load_fermentation_test_exception():
+    # Arrange
+
+    # Act and Assert
+    with pytest.raises(ValueError):
+        test_spectra, test_hplc = load_fermentation_test(set_output="plars")
+
+
+def test_load_fermentation_train_pandas():
     # Arrange
 
     # Act
@@ -40,4 +89,23 @@ def test_load_fermentation_train():
     assert isinstance(train_spectra, pd.DataFrame)
     assert isinstance(train_hplc, pd.DataFrame)
 
-    
+
+def test_load_fermentation_train_polars():
+    # Arrange
+
+    # Act
+    train_spectra, train_hplc = load_fermentation_train(set_output="polars")
+
+    # Assert
+    assert train_spectra.shape == (21, 1047)
+    assert train_hplc.shape == (21, 1)
+    assert isinstance(train_spectra, pl.DataFrame)
+    assert isinstance(train_hplc, pl.DataFrame)
+
+
+def test_load_fermentation_train_exception():
+    # Arrange
+
+    # Act and Assert
+    with pytest.raises(ValueError):
+        train_spectra, train_hplc = load_fermentation_train(set_output="plars")
