+++ chemotools/datasets/_base.py
@@ -1,13 +1,21 @@
-import pandas as pd
 import os
+
+
+import pandas as pd
+import polars as pl
 
 PACKAGE_DIRECTORY = os.path.dirname(os.path.abspath(__file__))
 
 
-def load_fermentation_train():
+def load_fermentation_train(set_output="pandas"):
     """
-    Loads the training data of the fermentation dataset. This data corresponds to a synthetic dataset measured 
+    Loads the training data of the fermentation dataset. This data corresponds to a synthetic dataset measured
     off-line. This dataset is designed to represent the variability of real fermentation data.
+
+    Arguments
+    -------
+    set_output: str, default='pandas'
+        The output format of the data. It can be 'pandas' or 'polars'. If 'polars', the data is returned as a polars DataFrame.
 
     Returns
     -------
@@ -20,17 +28,32 @@ def load_fermentation_train():
     Mauricio Iglesias Miguel, Gernaey Krist V. Transforming data into information:
     A parallel hybrid model for real-time state estimation in lignocellulose ethanol fermentations.
     """
-    train_spectra = pd.read_csv(PACKAGE_DIRECTORY + "/data/train_spectra.csv")
-    train_spectra.columns = train_spectra.columns.astype(float)
-    train_hplc = pd.read_csv(PACKAGE_DIRECTORY + "/data/train_hplc.csv")
+    if set_output == "pandas":
+        train_spectra = pd.read_csv(PACKAGE_DIRECTORY + "/data/train_spectra.csv")
+        train_spectra.columns = train_spectra.columns.astype(float)
+        train_hplc = pd.read_csv(PACKAGE_DIRECTORY + "/data/train_hplc.csv")
+        return train_spectra, train_hplc
 
-    return train_spectra, train_hplc
+    if set_output == "polars":
+        train_spectra = pl.read_csv(PACKAGE_DIRECTORY + "/data/train_spectra.csv")
+        train_hplc = pl.read_csv(PACKAGE_DIRECTORY + "/data/train_hplc.csv")
+        return train_spectra, train_hplc
+
+    else:
+        raise ValueError(
+            "Invalid value for set_output. Please use 'pandas' or 'polars'."
+        )
 
 
-def load_fermentation_test():
+def load_fermentation_test(set_output="pandas"):
     """
     Loads the testing data of the fermentation dataset. This data corresponds to real fermentation data measured
-    on-line during a fermentation process. 
+    on-line during a fermentation process.
+
+    Arguments
+    -------
+    set_output: str, default='pandas'
+        The output format of the data. It can be 'pandas' or 'polars'. If 'polars', the data is returned as a polars DataFrame.
 
     Returns
     -------
@@ -43,27 +66,57 @@ def load_fermentation_test():
     Mauricio Iglesias Miguel, Gernaey Krist V. Transforming data into information:
     A parallel hybrid model for real-time state estimation in lignocellulose ethanol fermentations.
     """
-    fermentation_spectra = pd.read_csv(
-        PACKAGE_DIRECTORY + "/data/fermentation_spectra.csv"
-    )
-    fermentation_spectra.columns = fermentation_spectra.columns.astype(float)
-    fermentation_hplc = pd.read_csv(PACKAGE_DIRECTORY + "/data/fermentation_hplc.csv")
+    if set_output == "pandas":
+        fermentation_spectra = pd.read_csv(
+            PACKAGE_DIRECTORY + "/data/fermentation_spectra.csv"
+        )
+        fermentation_spectra.columns = fermentation_spectra.columns.astype(float)
+        fermentation_hplc = pd.read_csv(
[…]
