### 2024-02-11 a50b353
update requirements.txt to provide polars support
