+++ requirements.txt
@@ -1,3 +1,5 @@
 numpy>=1.24.1
 pandas>=1.3.4
-scikit-learn>=1.2.0 
+polars>=0.20.0
+pyarrow>=15.0.0
+scikit-learn>=1.4.0 
