+++ releases/0.1.5.md
@@ -1,0 +1,15 @@
+---
+title: v0.1.4
+layout: default
+parent: Releases
+---
+
+# __[Release v0.1.4](https://github.com/paucablop/chemotools/releases/tag/v0.1.4)__
+
+## __What's new? 🎉🎉__
+
+## __Improvements ✨✨__
+
+Polars support for datasets and for all functions. Now you can load the datasets as polars.DataFrame and use the functions with polars.DataFrame.
+
+## __Bug fixes 🐛🐛__
