### 2024-02-12 4dcd65c
fix version issue in 0.1.5 release documentation
