+++ releases/0.1.5.md
@@ -1,10 +1,10 @@
 ---
-title: v0.1.4
+title: v0.1.5
 layout: default
 parent: Releases
 ---
 
-# __[Release v0.1.4](https://github.com/paucablop/chemotools/releases/tag/v0.1.4)__
+# __[Release v0.1.5](https://github.com/paucablop/chemotools/releases/tag/v0.1.5)__
 
 ## __What's new? 🎉🎉__
 
