+++ releases/0.1.5.md
@@ -10,6 +10,6 @@ parent: Releases
 
 ## __Improvements ✨✨__
 
-Polars support for datasets and for all functions. Now you can load the datasets as polars.DataFrame and use the functions with polars.DataFrame.
+Polars support for datasets and for all functions. Now you can load the datasets as ```polars.DataFrame``` and use the functions with ```polars.DataFrame```.
 
 ## __Bug fixes 🐛🐛__
