### 2024-04-20 11f4937
feat: re-added improved version of banded LU-decomposition; added tests
