+++ tests/test_for_utils.py
@@ -2,10 +2,15 @@ from typing import Literal
 
 import numpy as np
 import pytest
-from scipy.linalg import cholesky_banded, eigvals_banded, solveh_banded
+from scipy.linalg import cholesky_banded, eigvals_banded, solve_banded, solveh_banded
 from scipy.sparse import eye as speye
 
-from chemotools.utils.banded_linalg import conv_to_lu_banded_storage, slodget_cho_banded
+from chemotools.utils.banded_linalg import (
+    conv_to_lu_banded_storage,
+    lu_banded,
+    lu_solve_banded,
+    slodget_cho_banded,
+)
 from chemotools.utils.finite_differences import (
     calc_forward_diff_kernel,
     calc_limit_max_eigval_fin_diff_mat,
@@ -78,6 +83,103 @@ def test_forward_finite_diff_conv_matrix(size: int, difference: int) -> None:
         f"Differences by matrix product for difference order {difference} with "
         f"accuracy 1 for series of size {size} failed."
     )
+
+
+@pytest.mark.parametrize("with_finite_check", [True, False])
+@pytest.mark.parametrize("overwrite_b", [True, False])
+@pytest.mark.parametrize("n_rhs", [0, 1, 2])
+@pytest.mark.parametrize("n_upp_bands", [1, 2, 3, 4, 5, 6])
+@pytest.mark.parametrize("n_low_bands", [1, 2, 3, 4, 5, 6])
+@pytest.mark.parametrize(
+    "n_rows", [1, 2, 10, 11, 50, 51, 100, 101, 500, 501, 1_000, 1001, 5_000, 5001]
+)
+def test_lu_banded_solve(
+    n_rows: int,
+    n_low_bands: int,
+    n_upp_bands: int,
+    n_rhs: int,
+    overwrite_b: bool,
+    with_finite_check: bool,
+) -> None:
+    """
+    Tests the separate LU decomposition followed by solving a system of linear equations
+    for banded matrices.
+
+    NOTE: A number of 0 right-hand sides are used for making the vector to solve a
+    NOTE: 1D-Array.
+
+    """
+
+    # if the matrix cannot exist with the given shape, the test is skipped
+    n_rows_min = n_low_bands + n_upp_bands + 1
+    if n_rows < n_rows_min:
+        pytest.skip(
+            f"Test skipped because the number of rows {n_rows} is smaller than the "
+            f"minimum number of rows {n_rows_min} required by the number of sub- "
+            f"{n_low_bands} and superdiagonals {n_upp_bands}."
+        )
+
+    # a random banded matrix and right-hand-side-vector/-matrix are generated
+    np.random.seed(seed=42)
+    ab = np.random.rand(n_low_bands + n_upp_bands + 1, n_rows)
+    b = np.random.rand(n_rows) if n_rhs == 0 else np.random.rand(n_rows, n_rhs)
+
+    # first, the Scipy solution is computed because if this fails due to singularity,
+    # the test has to not test for equivalent results, but for failure
+    # NOTE: failure is indicated by the solution being ``None``
+    # NOTE: this order of evaluation is also better for testing if the overwrite flag
+    # NOTE: is working correctly because otherwise SciPy would get the overwritten b
+    l_and_u = (n_low_bands, n_upp_bands)
+    x_ref = None
+    try:
+        x_ref = solve_banded(
+            l_and_u=l_and_u,
+            ab=ab,
+            b=b,
+            check_finite=True,
+        )
+
+        # NOTE: even if SciPy computes the solution "successfully", there might be NaNs
+        # NOTE: in the result, so the test has to check for that as well
+        if np.any(np.isnan(x_ref)):
+            x_ref = None
+
+    except np.linalg.LinAlgError:
+        pass
+
[…]
