+++ chemotools/utils/banded_linalg.py
@@ -1,8 +1,30 @@
 from numbers import Integral
 
 import numpy as np
+from numpy.typing import ArrayLike
+from scipy.linalg import lapack
 from scipy.sparse import spmatrix
 from sklearn.utils import check_array, check_scalar
+
+from chemotools.utils.models import BandedLUFactorization
+
+
+def _datacopied(arr, original):
+    """
+    Strictly check for ``arr`` not sharing any data wit ``original``, under the
+    assumption that ``arr = asarray(original)``
+
+    Was copied from Scipy to be consistent in the LAPACK-wrappers implemented here.
+
+    """
+
+    if arr is original:
+        return False
+
+    if not isinstance(original, np.ndarray) and hasattr(original, "__array__"):
+        return False
+
+    return arr.base is None
 
 
 def _check_full_arr_n_diag_counts_for_lu_banded(
@@ -146,6 +168,222 @@ def conv_to_lu_banded_storage(
         ab[main_diag_idx - offset, 0:offset] = diag_method(offset)
 
     return ab
+
+
+def lu_banded(
+    l_and_u: tuple[int, int],
+    ab: ArrayLike,
+    *,
+    check_finite: bool = True,
+) -> BandedLUFactorization:
+    """
+    Computes the LU-decomposition of a banded matrix ``A`` using LAPACK-routines.
+    This function is a wrapper of the LAPACK-routine ``gbtrf`` which computes the LU-
+    decomposition of a banded matrix ``A`` in-place. It wraps the routine in an
+    analogous way to SciPy's ``scipy.linalg.cholesky_banded``.
+
+    Parameters
+    ----------
+    l_and_u : (int, int)
+        The number of "non-zero" sub- (first) and superdiagonals (second element) aside
+        the main diagonal which does not need to be considered here. "Non-zero" can be
+        a bit misleading in this context. These numbers should count up to the diagonal
+        after which all following diagonals are all zero. Zero-diagonals that come
+        before still need to be included.
+        Neither of both may exceed ``num_rows``.
+        Wrong specification of this can lead to non-zero-diagonals being ignored or
+        zero-diagonals being included which corrupts the results or reduces the
+        performance.
+    ab : array_like of shape (l_and_u[0] + 1 + l_and_u[1], n)
+        A 2D-Array resembling the matrix ``A`` in banded storage format (see Notes).
+    check_finite : bool, default=True
+        Whether to check that the input matrix contains only finite numbers. Disabling
+        may give a performance gain, but may result in problems (crashes,
+        non-termination) if the inputs do contain infinities or NaNs.
+
+    Returns
+    -------
+    lub_factorization : BandedLUFactorization
+        A dataclass containing the LU-factorization of the matrix ``A`` as follows:
+            ``lub``: The LU-decomposition of ``A`` in banded storage format (see Notes).
+            ``ipiv``: The pivoting indices.
+            ``l_and_u``: The number of sub- and superdiagonals of the matrix ``A`` that
+                are non-zero.
+            ``singular``: A boolean indicating whether the matrix is singular.
+
+    Notes
+    -----
+    For LAPACK LU-decomposition, the matrix ``a`` is stored in ``ab`` using the matrix
+    diagonal ordered form:
+
+        ```python
+        ab[u + i - j, j] == a[i,j] # see below for u
+        ```
+
+    An example of ``ab`` (shape of a is ``(7,7)``, ``u``=3 superdiagonals, ``l``=2
+    subdiagonals) looks like:
+
+        ```python
+             *    *    *   a03  a14  a25  a36
+             *    *   a02  a13  a24  a35  a46
[…]
