### 2024-04-20 556b92f
feat: added wrappers for a dataclass-based banded Cholesky decomposition and solve; added preliminary tests
