+++ chemotools/utils/banded_linalg.py
@@ -2,11 +2,13 @@ from numbers import Integral
 
 import numpy as np
 from numpy.typing import ArrayLike
+from scipy.linalg import cho_solve_banded as scipy_cho_solve_banded
+from scipy.linalg import cholesky_banded as scipy_cholesky_banded
 from scipy.linalg import lapack
 from scipy.sparse import spmatrix
 from sklearn.utils import check_array, check_scalar
 
-from chemotools.utils.models import BandedLUFactorization
+from chemotools.utils.models import BandedCholeskyFactorization, BandedLUFactorization
 
 
 def _datacopied(arr, original):
@@ -349,9 +351,9 @@ def lu_solve_banded(
 
     # then, the shapes of the LU-decomposition and ``b`` need to be validated against
     # each other
-    if lub_factorization.shape[-1] != b_inter.shape[0]:
+    if lub_factorization.n_cols != b_inter.shape[0]:
         raise ValueError(
-            f"\nShapes of lub ({lub_factorization.shape[-1]}) and b "
+            f"\nShapes of lub ({lub_factorization.n_cols}) and b "
             f"({b_inter.shape[0]}) are not compatible."
         )
 
@@ -386,7 +388,143 @@ def lu_solve_banded(
     )
 
 
-def slodget_cho_banded(decomposition: tuple[np.ndarray, bool]) -> tuple[float, float]:
+def slogdet_lu_banded(
+    lub_factorization: BandedLUFactorization,
+) -> tuple[float, float]:
+    """
+    Computes the logarithm of the absolute value and the sign of the determinant of a
+    banded matrix A using its LU-decomposition. This is way more efficient than
+    computing the determinant directly because the LU-decompositions main diagonals
+    already encode the determinant as the product of the diagonal entries of the
+    factors.
+
+    Parameters
+    ----------
+    lub_factorization : BandedLUFactorization
+        The LU-decomposition of the matrix ``A`` in banded storage format as returned by
+        the function :func:`lu_banded`.
+
+    Returns
+    -------
+    sign : float
+        A number representing the sign of the determinant.
+    logabsdet : float
+        The natural log of the absolute value of the determinant.
+        If the determinant is zero, then `sign` will be 0 and `logabsdet` will be
+        -Inf. In all cases, the determinant is equal to ``sign * np.exp(logabsdet)``.
+
+    Raises
+    ------
+    OverflowError
+        If any of the diagonal entries of the LU-decomposition leads to an overflow in
+        the natural logarithm.
+
+    """
+
+    # first, the number of actual row exchanges needs to be counted
+    unchanged_row_idxs = np.arange(
+        start=0,
+        stop=lub_factorization.n_rows,
+        step=1,
+        dtype=lub_factorization.ipiv.dtype,
+    )
+    num_row_exchanges = np.count_nonzero(lub_factorization.ipiv - unchanged_row_idxs)
+
+    # the sign-prefactor of the determinant is either +1 or -1 depending on whether the
+    # number of row exchanges is even or odd
+    sign = -1.0 if num_row_exchanges % 2 == 1 else 1.0
+
+    # since the determinant (without sign prefactor) is just the product of the diagonal
+    # product of L and the diagonal product of U, the calculation simplifies. As the
+    # main diagonal of L is a vector of ones, only the diagonal product of U is required
+    main_diag = lub_factorization.lub[lub_factorization.main_diag_row_idx, ::]
+    u_diaprod_sign = np.prod(np.sign(main_diag))
+    with np.errstate(divide="ignore", over="ignore"):
+        logabsdet = np.sum(np.log(np.abs(main_diag)))
+
+    # logarithms of zero are already properly handled, so there is not reason to worry
+    # about, since they are -inf which will result in a zero determinant in exp();
[…]
