### 2024-04-21 064afcf
refactor: implemented fast computation of both squared forward finite differences matrix version that does not rely on `scipy.sparse` anymore to avoid overhead and redundant computations
