+++ chemotools/utils/finite_differences.py
@@ -2,9 +2,6 @@ from math import comb
 from numbers import Integral
 
 import numpy as np
-from scipy.sparse import csr_matrix, dia_matrix
-from scipy.sparse import diags as spdiags
-from scipy.sparse import eye as speye
 from sklearn.utils import check_scalar
 
 
@@ -12,7 +9,8 @@ def calc_forward_diff_kernel(
     *,
     differences: int,
 ) -> np.ndarray:
-    """Computes the kernel for forward finite differences which can be applied to a
+    """
+    Computes the kernel for forward finite differences which can be applied to a
     series by means of a convolution, e.g.,
 
     ```python
@@ -27,18 +25,18 @@ def calc_forward_diff_kernel(
         The order of the differences starting from 0 for the original curve, 1 for the
         first order, 2 for the second order, ..., and ``m`` for the ``m``-th order
         differences.
-        Values below 0 are not allowed.
+        Values below 1 are not allowed.
 
     Returns
     -------
-    fin_diff_kernel : np.ndarray of shape (differences + 1,)
+    fin_diff_kernel : ndarray of shape (differences + 1,)
         A NumPy-1D-vector resembling the kernel from the code example above. To avoid
         loss of precision, the data type is ``np.int64``.
 
     Raises
     ------
     ValueError
-        If the difference order is below 0.
+        If the difference order is below 1.
 
     """
     # the input is validated
@@ -46,7 +44,7 @@ def calc_forward_diff_kernel(
         differences,
         name="differences",
         target_type=Integral,
-        min_val=0,
+        min_val=1,
         include_boundaries="left",
     )
 
@@ -60,192 +58,202 @@ def calc_forward_diff_kernel(
     )
 
 
-def forward_finite_diff_conv_matrix(
+def _gen_squ_fw_fin_diff_mat_cho_banded_transp_first(
     *,
+    n_data: int,
     differences: int,
-    series_size: int,
-) -> dia_matrix:
-    """Computes the convolution matrix for forward finite differences which can be
-    applied to a series by means of a matrix multiplication, e.g.,
+) -> np.ndarray:
+    """
+    Generates the squared forward finite differences matrix ``D.T @ D`` from the
+    forward finite difference matrix ``D`` of order ``differences``. It will be cast to
+    to the upper banded storage format used for LAPACK's banded Cholesky decomposition.
 
-    ```python
-        conv_mat = finite_diff_conv_matrix(differences=2, series_size=10)
-        differences = conv_mat @ series # boundaries require special care
-    ```
-
-    this is equivalent to
-
-    ```python
-        kernel = calc_forward_fin_diff_kernel(differences=2)
-        differences = np.convolve(series, np.flip(kernel), mode="valid")
-        # NOTE: NumPy flips the kernel internally due to the definition of convolution
-    ```
-
-    Parameters
-    ----------
-    differences : int
-        The order of the differences starting from 0 for the original curve, 1 for the
-        first order, 2 for the second order, ..., and ``m`` for the ``m``-th order
-        differences.
-        Values below 0 are not allowed.
-    series_size : int
-        The number of data points in the series to which the convolution matrix is
-        applied.
-
-    Returns
-    -------
-    diff_mat : dia_matrix of shape (series_size - differences, series_size)
-        A sparse matrix resembling the convolution matrix from the code example above.
-        To avoid loss of precision, the data type is ``np.int64``.
-
-    Raises
-    ------
-    ValueError
-        If the difference order is below 0, or ``series_size`` is not sufficient to
-        support the respective difference order.
[…]
