### 2024-04-22 097d4f0
refactor: renamed decomposition model to solver model; added PentaPy model
