### 2024-04-22 bc2198b
feat: added new banded storage conversion function; added submodule header; added sections; renamed "LU-" to "LU "
