+++ chemotools/utils/banded_linalg.py
@@ -1,3 +1,22 @@
+"""
+This utility submodule provides functions for the linear algebra with banded matrices,
+namely
+
+- conversion from the upper banded storage for LAPACK's banded Cholesky decomposition
+    to the banded storage for LAPACK's banded LU decomposition,
+- wrappers for SciPy's LAPACK-routines for the banded Cholesky decomposition and the
+    corresponding linear solver,
+- LU decomposition of a banded matrix and the corresponding linear solver,
+- computation of the log-determinant of a banded matrix using its Cholesky or LU
+    decomposition
+
+The decomposition functions return dataclasses that facilitate the handling of the
+factorizations.
+
+"""
+
+### Imports ###
+
 from numbers import Integral
 
 import numpy as np
@@ -9,6 +28,13 @@ from scipy.sparse import spmatrix
 from sklearn.utils import check_array, check_scalar
 
 from chemotools.utils.models import BandedCholeskyFactorization, BandedLUFactorization
+
+### Type Aliases ###
+
+LAndUBandCounts = tuple[int, int]
+
+
+### Auxiliary Functions ###
 
 
 def _datacopied(arr, original):
@@ -34,7 +60,7 @@ def _check_full_arr_n_diag_counts_for_lu_banded(
     l_and_u: tuple[int, int],
 ) -> None:
     """Validates the shape of a full array and the number of sub- and superdiagonals
-    for LU-decomposition of a banded (sparse) matrix.
+    for LU decomposition of a banded (sparse) matrix.
     """
     num_rows, num_cols = a_shape
     num_low_diags, num_upp_diags = l_and_u
@@ -75,12 +101,99 @@ def _check_full_arr_n_diag_counts_for_lu_banded(
     # else nothing
 
 
+def conv_upper_chol_banded_to_lu_banded_storage(
+    ab: np.ndarray,
+) -> tuple[LAndUBandCounts, np.ndarray]:
+    """
+    Converts the upper banded storage format used by LAPACK's banded Cholesky
+    decomposition to the banded storage format used by LAPACK's banded LU
+    decomposition.
+
+    Parameters
+    ----------
+    ab : np.ndarray of shape (n_upp_bands + 1, n_cols)
+        The matrix ``A`` stored in the upper banded storage format used by LAPACK's
+        banded Cholesky decomposition (see Notes for details).
+
+    Returns
+    -------
+    l_and_u : (int, int)
+        The number of sub- (first) and superdiagonals (second element) aside the main
+        diagonal which does not need to be considered here.
+    ab : np.ndarray of shape (l_and_u[0] + 1 + l_and_u[1], n_cols)
+        The matrix ``A`` stored in the banded storage format used by LAPACK's banded LU
+        decomposition (see Notes for details).
+
+    Notes
+    -----
+    The upper diagonal ordered form for LAPACK's Cholesky decomposition is given by the
+    following ordering
+
+    ```python
+    ab[u + i - j, j] == a[i,j]
+    ```
+
+    e.g., for a symmetric matrix ``A`` of shape (7, 7) with in total 3 superdiagonals,
+    3 subdiagonals, and the main diagonal, the ordering is as follows:
+
+    ```python
+    *   *   *   a03 a14 a25 a36
+    *   *   a02 a13 a24 a35 a46
+    *   a01 a12 a23 a34 a45 a56 # ^ superdiagonals
+    a00 a11 a22 a33 a44 a55 a66 # main diagonal
+    ```
+
+    where each `*` denotes a zero element.
+
+    For LAPACK's LU decomposition, the matrix `A` is stored in `ab` using the matrix
+    diagonal ordered form:
+
+    ```python
+    ab[u + i - j, j] == a[i,j]
+    ```
+
+    The example from above would then look like this where basicall, all the
[…]
