+++ chemotools/utils/banded_linalg.py
@@ -4,11 +4,8 @@ namely
 
 - conversion from the upper banded storage for LAPACK's banded Cholesky decomposition
     to the banded storage for LAPACK's banded LU decomposition,
-- wrappers for SciPy's LAPACK-routines for the banded Cholesky decomposition and the
-    corresponding linear solver,
 - LU decomposition of a banded matrix and the corresponding linear solver,
-- computation of the log-determinant of a banded matrix using its Cholesky or LU
-    decomposition
+- computation of the log-determinant of a banded matrix using its LU decomposition
 
 The decomposition functions return dataclasses that facilitate the handling of the
 factorizations.
@@ -21,13 +18,11 @@ from numbers import Integral
 
 import numpy as np
 from numpy.typing import ArrayLike
-from scipy.linalg import cho_solve_banded as scipy_cho_solve_banded
-from scipy.linalg import cholesky_banded as scipy_cholesky_banded
 from scipy.linalg import lapack
 from scipy.sparse import spmatrix
 from sklearn.utils import check_array, check_scalar
 
-from chemotools.utils.models import BandedCholeskyFactorization, BandedLUFactorization
+from chemotools.utils.models import BandedLUFactorization
 
 ### Type Aliases ###
 
@@ -152,7 +147,7 @@ def conv_upper_chol_banded_to_lu_banded_storage(
     ab[u + i - j, j] == a[i,j]
     ```
 
-    The example from above would then look like this where basicall, all the
+    The example from above would then look like this where basically, all the
     superdiagonal rows are just copied to the subdiagonal rows and moved to the left so
     that the first non-zero element of each row is in the first column:
 
@@ -578,116 +573,3 @@ def slogdet_lu_banded(
         return sign, logabsdet
 
     return -sign, logabsdet
-
-
-### SciPy-Wrappers for banded Cholesky-decomposition ###
-
-
-def cholesky_banded(
-    ab: np.ndarray,
-    overwrite_ab: bool = False,
-    lower: bool = False,
-    check_finite: bool = True,
-) -> BandedCholeskyFactorization:
-    """
-    A drop-in replacement for SciPy's ``cholesky_banded`` that stores the factorization
-    in a dataclass.
-
-    Please refer to the SciPy documentation for further information that is not
-    mentioned here.
-
-    Returns
-    -------
-    chob_factorization : BandedCholeskyFactorization
-        A dataclass containing the Cholesky-factorization of the matrix ``A`` as
-        follows:
-            ``lb``: The Cholesky-decomposition of ``A`` in banded storage format.
-            ``lower``: A boolean indicating whether the Cholesky-decomposition is in
-                lower triangular form (``True``) or in upper triangular form
-                (``False``).
-    """
-
-    return BandedCholeskyFactorization(
-        lb=scipy_cholesky_banded(**locals()),
-        lower=lower,
-    )
-
-
-def cho_solve_banded(
-    chob_factorization: BandedCholeskyFactorization,
-    b: np.ndarray,
-    overwrite_b: bool = False,
-    check_finite: bool = True,
-) -> np.ndarray:
-    """
-    A drop-in replacement for SciPy's ``cho_solve_banded`` that relies on the
-    factorization being stored in a dataclass.
-
-    Please refer to the SciPy documentation for further information that is not
-    mentioned here.
-
-    Parameters
-    ----------
-    chob_factorization : BandedCholeskyFactorization
-        The Cholesky-factorization of the matrix ``A`` in banded storage format as
-        returned by the function :func:`cholesky_banded`.
-
-    """
-
-    return scipy_cho_solve_banded(
-        cb_and_lower=(chob_factorization.lb, chob_factorization.lower),
[…]
