### 2024-04-22 8cfa2f5
feat: added submodule header; organized submodule; added automated smoother names enum
