### 2024-04-22 f7f6638
refactor: major refactoring of Whittaker smoother base class; removed Cholesky; adapted models; used iterators; used more efficient numerical operations; incorporated fast pre-computations; got rid off sparse matrix representations via SciPy
