+++ chemotools/utils/whittaker_base.py
@@ -1,69 +1,94 @@
-import numpy as np
-from scipy.linalg import cho_solve_banded, cholesky_banded
-from scipy.sparse import csr_matrix, dia_matrix
+"""
+This utility submodule provides the base class for the Whittaker-like smoothing
+algorithm. It is used to solve linear systems of equations that involve banded
+matrices as they occur in applications like the Whittaker-Henderson-smoothing or
+derived methods like Asymmetric Least Squares (ALS) baseline correction.
 
-from chemotools.utils.banded_linalg import conv_to_lu_banded_storage
-from chemotools.utils.finite_differences import (
-    forward_finite_diff_conv_matrix,
-    posdef_mod_squared_fw_fin_diff_conv_matrix,
+"""
+
+### Imports ###
+
+from typing import Generator, Optional, Union, overload
+
+import numpy as np
+
+from chemotools.utils.banded_linalg import (
+    LAndUBandCounts,
+    conv_upper_chol_banded_to_lu_banded_storage,
+    lu_banded,
+    lu_solve_banded,
+    slogdet_lu_banded,
 )
-from chemotools.utils.models import _PENTAPY_AVAILABLE, BandedSolveDecompositions
+from chemotools.utils.finite_differences import gen_squ_fw_fin_diff_mat_cho_banded
+from chemotools.utils.models import (
+    _PENTAPY_AVAILABLE,
+    AutoSmoothMethods,
+    BandedLUFactorization,
+    BandedPentapyFactorization,
+    BandedSolvers,
+)
 
 if _PENTAPY_AVAILABLE:
     import pentapy as pp
-# else nothing
 
-_CholeskyDecomposition = tuple[np.ndarray, bool]
-_PentapyDecomposition = None
-_Decomposition = _PentapyDecomposition | _CholeskyDecomposition
+### Type Aliases ###
+
+_Decomposition = Union[BandedLUFactorization, BandedPentapyFactorization]
+
+
+### Class Implementation ###
 
 
 class WhittakerLikeSolver:
-    """This class can be used to solve linear systems of equations that involve banded
+    """
+    This class can be used to solve linear systems of equations that involve banded
     matrices as they occur in applications like the Whittaker-Henderson-smoothing or
     derived methods like Asymmetric Least Squares (ALS) baseline correction.
     It support weights and tries to use the most efficient method available.
 
     Attributes
     ----------
-    series_size_ : int
-        The size of the series to smooth. It is equivalent to `n_features_in_`, but it
-        was renamed to be allow for definition after the initialisation.
-    lam_ : int or float
-        The lambda parameter to use for the Whittaker smooth.
+    n_data_ : int
+        The number of data points within the series to smooth. It is equivalent to
+        ``n_features_in_``, but it was renamed to be allow for definition after the
+        initialisation.
+    lam_ : int or float or AutoSmoothMethods
+        The lambda parameter to use for the smoothing, a.k.a. the penalty weight or
+        smoothing parameter.
+        If a member of :class:`AutoSmoothMethods` is provided, the lambda parameter is
+        fitted automatically, but then the pre-computations in meth:`_setup_for_fit`
+        and/or :meth:`_whittaker_solve` might take significantly longer because more
+        pre-computations are required and multiple penalty weights are tested.
     differences_ : int
-        The number of differences to use for the Whittaker smooth. If the aim is to
-        obtain a smooth estimate of the ``m``-th order derivative, this should be set to
+        The number of differences to use for the smoothing. If the aim is to obtain a
+        smooth estimate of the ``m``-th order derivative, this should be set to
         at least ``m + 2``.
-    l_and_u_ : tuple[int, int]
-        The number of sub- (first) and superdiagonals (second element). Both will equal
-        ``differences_``.
-    fw_fin_diff_mat_ : dia_matrix
[…]
