### 2024-04-22 6098848
refactor: went for more secure way of specifying the penalty weight
