+++ chemotools/utils/models.py
@@ -8,6 +8,7 @@ dataclasses used throughout the package.
 
 from dataclasses import dataclass, field
 from enum import Enum
+from typing import Union
 
 import numpy as np
 
@@ -38,7 +39,7 @@ class BandedSolvers(str, Enum):
 
     """
 
-    PIVOTED_LU = "pivoted LU decomposition"
+    PIVOTED_LU = "partially pivoted LU decomposition"
     PENTAPY = "direct pentadiagonal solver"
 
 
@@ -62,6 +63,70 @@ class AutoSmoothMethods(str, Enum):
 ### (Data) Classes ###
 
 
+# a dataclass for specification of the smoothing penalty weight lambda for the
+# Whittaker-Henderson smoother
+
+
+@dataclass()
+class WhittakerSmoothLambda:
+    """
+    A dataclass that holds the specification of the smoothing penalty weight or
+    smoothing parameter lambda for the Whittaker-Henderson smoother.
+
+    Attributes
+    ----------
+    low_bound, upp_bound: int or float
+        The lower and upper bound of the search space for the penalty weight.
+        Flipped bounds are automatically corrected, but they have to differ by at least
+        a factor of 10.
+    method: AutoSmoothMethods
+        The method to use for the automatic selection of the penalty weight.
+
+    Raises
+    ------
+    ValueError
+        If ``upp_bound`` is not greater than 10 times ``low_bound`` after eventually
+        flipping the bounds.
+
+    """
+
+    low_bound: Union[int, float]
+    upp_bound: Union[int, float]
+    method: AutoSmoothMethods
+
+    def __post_init__(self):
+        # firs, the input types are checked
+        if not isinstance(self.low_bound, (int, float)) or not isinstance(
+            self.upp_bound, (int, float)
+        ):
+            raise TypeError(
+                f"\nThe lower bound ({self.low_bound}) and upper bound "
+                f"({self.upp_bound}) have to be integers or floats."
+            )
+
+        if not isinstance(self.method, AutoSmoothMethods):
+            raise TypeError(
+                f"\nThe method ({self.method}) has to be a member of the "
+                f"AutoSmoothMethods."
+            )
+
+        # then, the lower and upper bound are sanitized by swapping them if necessary
+        # and checking if the upper bound is at least 10 times the lower bound
+        if self.low_bound >= self.upp_bound:
+            self.low_bound, self.upp_bound = self.upp_bound, self.low_bound
+
+        if self.upp_bound < 10 * self.low_bound:
+            raise ValueError(
+                f"\nThe upper bound ({self.upp_bound}) has to be at least 10 times the "
+                f"lower bound ({self.low_bound})."
+            )
+
+
+# a fake class for representing the factorization of a pentadiagonal matrix with
+# pentapy which is empty since pentapy does not factorize the matrix but directly solves
+# the system of equations
+
+
 class BandedPentapyFactorization:
     """
     A class that resembles the factorization of a pentadiagonal matrix with ``pentapy``.
@@ -71,6 +136,10 @@ class BandedPentapyFactorization:
     """
 
     pass
+
+
+# a dataclass for the factorization of a banded matrix with LU decomposition with p
+# partial pivoting
 
 
 @dataclass()
