### 2024-04-22 254f906
refactor: replaced intermediate solver distributed method by a dictionary to remove one indirection with one branched logic
