### 2024-05-01 7b5385f
style: went for using `ndarray`-methods
