### 2024-05-01 51261d3
feat: added fixed lambda to models
