+++ chemotools/utils/models.py
@@ -8,7 +8,8 @@ dataclasses used throughout the package.
 
 from dataclasses import dataclass, field
 from enum import Enum
-from typing import Union
+from math import log
+from typing import Literal, Optional, Tuple, Union
 
 import numpy as np
 
@@ -43,22 +44,30 @@ class BandedSolvers(str, Enum):
     PENTAPY = "direct pentadiagonal solver"
 
 
-# an Enum class for the kinds of automated smoothing by the Whittaker-Henderson smoother
-# that can be applied to the data
+# an Enum class for the kinds of smoothing by the Whittaker-Henderson smoother that can
+# be applied to the data
 
 
-class AutoSmoothMethods(str, Enum):
+class WhittakerSmoothMethods(str, Enum):
     """
-    Defines the types of automated smoothing methods that can be applied to the data
-    using the Whittaker-Henderson smoother, i.e.,
+    Defines the types of smoothing methods that can be applied to the data using the
+    Whittaker-Henderson smoother, i.e.,
 
-    - ``LOG_MARGINAL_LIKELIHOOD``: smoothing based on the maximization of the log
-        marginal likelihood
+    - ``FIXED``: fixed penalty weight (shorthand "fixed")
+    - ``LOGML``: smoothing based on the maximization of the log marginal likelihood
+        (shorthand "logml")
+
+    Except for ``FIXED``, the penalty weight is automatically determined when using the
+    other methods.
 
     """
 
-    LOG_MARGINAL_LIKELIHOOD = "log marginal likelihood"
+    FIXED = "fixed"
+    LOGML = "logml"
 
+
+# a type hint is defined for the Whittaker-Henderson smoother specification
+_WhittakerSmoothMethodsAll = Union[WhittakerSmoothMethods, Literal["fixed", "logml"]]
 
 ### (Data) Classes ###
 
@@ -75,51 +84,148 @@ class WhittakerSmoothLambda:
 
     Attributes
     ----------
-    low_bound, upp_bound: int or float
-        The lower and upper bound of the search space for the penalty weight.
-        Flipped bounds are automatically corrected, but they have to differ by at least
-        a factor of 10.
-    method: AutoSmoothMethods
-        The method to use for the automatic selection of the penalty weight.
+    bounds: int or float or (int or float, int or float)
+        The bounds for the search space of the penalty weight lambda. The specification
+        can be either
+
+        - a single value for a fixed penalty weight (requires ``method`` to be set to
+            ``WhittakerSmoothMethods.FIXED``), or
+        - a tuple of two values for the lower and upper bounds of the search space
+            (then ``method`` may not be set to ``WhittakerSmoothMethods.FIXED`` unless
+            the bounds are too close to each other as described below).
+
+        Independently of the specification, the values have to be greater than or equal
+        to the zero tolerance ``1e-25``.
+        If a lower and an upper bound are provided, they are flipped if necessary.
+        After that, the difference ``abs(upp_bound - low_bound)`` has to be at least
+        ``1e-5 * upp_bound`` for any method other than ``WhittakerSmoothMethods.FIXED``.
+        Otherwise, the method is set to ``WhittakerSmoothMethods.FIXED`` and the
+        ``fixed_lambda`` is set to the upper bound.
+    method: WhittakerSmoothMethods or {"fixed", "logml"}
+        The method to use for the selection of the penalty weight. If the bounds are too
+        close to each other, this will be set to ``WhittakerSmoothMethods.FIXED``.
 
     Raises
     ------
     ValueError
-        If ``upp_bound`` is not greater than 10 times ``low_bound`` after eventually
-        flipping the bounds.
+        If ``method`` is invalid, i.e., it does not correspond to any of the
[…]
