### 2024-05-01 a857ed2
refactor/feat: added automatic smoothing based on log marginal likelihood; re-formulated linear system for consistency with literature (without performance penalty); re-formulated weight handling
