+++ chemotools/utils/whittaker_base.py
@@ -8,9 +8,11 @@ derived methods like Asymmetric Least Squares (ALS) baseline correction.
 
 ### Imports ###
 
-from typing import Generator, Optional, Tuple, Union, overload
+from math import ceil, exp
+from typing import Generator, Optional, Tuple, Union
 
 import numpy as np
+from scipy.optimize import minimize_scalar
 
 from chemotools.utils.banded_linalg import (
     LAndUBandCounts,
@@ -19,14 +21,17 @@ from chemotools.utils.banded_linalg import (
     lu_solve_banded,
     slogdet_lu_banded,
 )
-from chemotools.utils.finite_differences import gen_squ_fw_fin_diff_mat_cho_banded
+from chemotools.utils.finite_differences import (
+    calc_forward_diff_kernel,
+    gen_squ_fw_fin_diff_mat_cho_banded,
+)
 from chemotools.utils.models import (
     _PENTAPY_AVAILABLE,
-    AutoSmoothMethods,
     BandedLUFactorization,
     BandedPentapyFactorization,
     BandedSolvers,
     WhittakerSmoothLambda,
+    WhittakerSmoothMethods,
 )
 
 if _PENTAPY_AVAILABLE:
@@ -34,11 +39,11 @@ if _PENTAPY_AVAILABLE:
 
 ### Type Aliases ###
 
-_Decomposition = Union[BandedLUFactorization, BandedPentapyFactorization]
+_Factorization = Union[BandedLUFactorization, BandedPentapyFactorization]
+_FactorizationForLogMarginalLikelihood = BandedLUFactorization
 _WhittakerSmoothLambdaPlain = Tuple[
-    Union[int, float], Union[int, float], AutoSmoothMethods
+    Union[int, float], Union[int, float], WhittakerSmoothMethods
 ]
-_LambdaInternal = Union[int, float, WhittakerSmoothLambda]
 
 
 ### Class Implementation ###
@@ -57,13 +62,6 @@ class WhittakerLikeSolver:
         The number of data points within the series to smooth. It is equivalent to
         ``n_features_in_``, but it was renamed to be allow for definition after the
         initialisation.
-    lam_ : int or float or WhittakerSmoothLambda
-        The lambda parameter to use for the smoothing, a.k.a. the penalty weight or
-        smoothing parameter.
-        If a member of :class:`WhittakerSmoothLambda` is provided, the lambda parameter
-        is fitted automatically, but then the pre-computations in meth:`_setup_for_fit`
-        and/or :meth:`_whittaker_solve` might take significantly longer because more
-        pre-computations are required and multiple penalty weights are tested.
     differences_ : int
         The number of differences to use for the smoothing. If the aim is to obtain a
         smooth estimate of the ``m``-th order derivative, this should be set to
@@ -71,12 +69,18 @@ class WhittakerLikeSolver:
         For higher orders, the systems to solve tend to get numerically instable,
         especially when ``n_data_`` grows large and high values for ``lam_`` are used.
         Values below 1 are not allowed.
-    _auto_fit_lam_ : bool
-        Whether the lambda parameter is fitted automatically (``True``) or fixed
-        (``False``).
+    _lam_inter_ : WhittakerSmoothLambda
+        The internal representation of the lambda parameter to use for the smoothing,
+        a.k.a. the penalty weight or smoothing parameter.
+        It is internally stored as an instance of the dataclass :class:`WhittakerSmoothLambda`.
     _l_and_u_ : (int, int)
         The number of sub- (first) and superdiagonals (second element) of the final
         matrix to solve for smoothing. Both elements will equal ``differences_``.
+    _diff_kernel_flipped_ : ndarray of shape (0, ) or (differences + 1,)
+        The flipped kernel to use for the forward finite differences. It is only
[…]
