### 2024-05-01 8569b6f
fix: fixed missing early returns
