### 2024-05-01 bdc7c34
fix: simplified weight checker; fixed broken multi-signal-single-weight-functionality
