### 2024-05-10 2664245
refactor: made log determinant sign computation less floating point dependent
