### 2024-05-10 c72de74
tests: added parallelizable, parametrized tests for utility functions; removed old utility tests
