+++ tests/test_for_utils/utils.py
@@ -1,0 +1,598 @@
+"""
+This script implements utility functions required for testing the
+:mod:`chemotools.utils` module.
+
+It contains doctests itself that are executed when running the script, but they are
+automatically tested when running the whole test suite as well. This ensures that the
+test utilities are working as expected as well.
+
+"""
+
+### Imports ###
+
+from typing import Tuple
+
+import numpy as np
+from scipy.sparse import csr_matrix
+from scipy.sparse import diags as sp_diags
+
+### Utility Functions ###
+
+
+def conv_upper_cho_banded_storage_to_sparse(
+    ab: np.ndarray,
+) -> csr_matrix:
+    """
+    Converts a banded matrix stored in the upper banded storage used for LAPACK's banded
+    Cholesky decomposition to a sparse ``CSR`` matrix.
+    For more information on the banded storage, please see the documentation of
+    :func:`chemotools.utils.banded_linalg.conv_upper_chol_banded_to_lu_banded_storage`.
+
+    Doctests
+    --------
+    >>> # Imports
+    >>> import numpy as np
+    >>> from numpy import nan
+    >>> from tests.test_for_utils.utils import conv_upper_cho_banded_storage_to_sparse
+
+    >>> # Generating a set of test matrices
+    >>> # Matrix 1
+    >>> ab = np.array(
+    ...     [
+    ...         [nan, nan,  1.,  2.,  3.],
+    ...         [nan,  4.,  5.,  6.,  7.],
+    ...         [ 8.,  9., 10., 11., 12.],
+    ...     ]
+    ... )
+    >>> conv_upper_cho_banded_storage_to_sparse(ab=ab).toarray()
+    array([[ 8.,  4.,  1.,  0.,  0.],
+           [ 4.,  9.,  5.,  2.,  0.],
+           [ 1.,  5., 10.,  6.,  3.],
+           [ 0.,  2.,  6., 11.,  7.],
+           [ 0.,  0.,  3.,  7., 12.]])
+
+    >>> # Matrix 2
+    >>> ab = np.array(
+    ...     [
+    ...         [nan, nan, nan,  1.],
+    ...         [nan, nan,  2.,  3.],
+    ...         [nan,  4.,  5.,  6.],
+    ...         [ 7.,  8.,  9., 10.],
+    ...     ]
+    ... )
+    >>> conv_upper_cho_banded_storage_to_sparse(ab=ab).toarray()
+    array([[ 7.,  4.,  2.,  1.],
+           [ 4.,  8.,  5.,  3.],
+           [ 2.,  5.,  9.,  6.],
+           [ 1.,  3.,  6., 10.]])
+
+    >>> # Matrix 3
+    >>> ab = np.array(
+    ...     [
+    ...         [1., 2., 3., 4., 5.],
+    ...     ]
+    ... )
+    >>> conv_upper_cho_banded_storage_to_sparse(ab=ab).toarray()
+    array([[1., 0., 0., 0., 0.],
+           [0., 2., 0., 0., 0.],
+           [0., 0., 3., 0., 0.],
+           [0., 0., 0., 4., 0.],
+           [0., 0., 0., 0., 5.]])
+
+    >>> # Matrix 4
+    >>> ab = np.array(
+    ...     [
+    ...         [nan,  1.],
+    ...         [ 2.,  3.],
+    ...     ]
+    ... )
+    >>> conv_upper_cho_banded_storage_to_sparse(ab=ab).toarray()
+    array([[2., 1.],
+           [1., 3.]])
+
+    >>> # Matrix 5
+    >>> ab = np.array(
+    ...     [
[…]
