+++ tests/test_for_utils/test_banded_linalg.py
@@ -1,0 +1,222 @@
+"""
+Test suite for the utility functions in the :mod:`chemotools.utils.banded_linalg`
+module.
+
+"""
+
+### Imports ###
+
+import numpy as np
+import pytest
+from scipy.linalg import eigvals_banded
+from scipy.linalg import solve_banded as scipy_solve_banded
+
+from chemotools.utils.banded_linalg import (
+    conv_upper_chol_banded_to_lu_banded_storage,
+    lu_banded,
+    lu_solve_banded,
+    slogdet_lu_banded,
+)
+
+### Test Suite ###
+
+
+@pytest.mark.parametrize("with_finite_check", [True, False])
+@pytest.mark.parametrize("overwrite_b", [True, False])
+@pytest.mark.parametrize("n_rhs", [0, 1, 2])
+@pytest.mark.parametrize("n_upp_bands", [1, 2, 3, 4, 5, 6])
+@pytest.mark.parametrize("n_low_bands", [1, 2, 3, 4, 5, 6])
+@pytest.mark.parametrize(
+    "n_rows",
+    [
+        1,
+        2,
+        3,
+        4,
+        5,
+        6,
+        7,
+        8,
+        9,
+        10,
+        11,
+        50,
+        51,
+        100,
+        101,
+        500,
+        501,
+        1_000,
+        1001,
+        5_000,
+        5001,
+    ],
+)
+def test_lu_banded_solve(
+    n_rows: int,
+    n_low_bands: int,
+    n_upp_bands: int,
+    n_rhs: int,
+    overwrite_b: bool,
+    with_finite_check: bool,
+) -> None:
+    """
+    Tests the separate LU decomposition followed by solving a system of linear equations
+    for banded matrices against the SciPy solution.
+
+    NOTE: A number of 0 right-hand sides are used for making the vector to solve a
+    NOTE: 1D-Array.
+
+    """
+
+    # if the matrix cannot exist with the given shape, the test is skipped
+    n_rows_min = n_low_bands + n_upp_bands + 1
+    if n_rows < n_rows_min:
+        pytest.skip(
+            f"Test skipped because the number of rows {n_rows} is smaller than the "
+            f"minimum number of rows {n_rows_min} required by the number of sub- "
+            f"{n_low_bands} and superdiagonals {n_upp_bands}."
+        )
+
+    # a random banded matrix and right-hand-side-vector/-matrix are generated
+    np.random.seed(seed=42)
+    ab = np.random.rand(n_low_bands + n_upp_bands + 1, n_rows)
+    b = np.random.rand(n_rows) if n_rhs == 0 else np.random.rand(n_rows, n_rhs)
+
+    # first, the Scipy solution is computed because if this fails due to singularity,
+    # the test has to not test for equivalent results, but for failure
+    # NOTE: failure is indicated by the solution being ``None``
+    # NOTE: this order of evaluation is also better for testing if the overwrite flag
+    #       is working correctly because otherwise SciPy would get the overwritten b
+    l_and_u = (n_low_bands, n_upp_bands)
+    x_ref = None
+    try:
+        x_ref = scipy_solve_banded(
+            l_and_u=l_and_u,
+            ab=ab,
+            b=b,
+            check_finite=True,
+        )
+
+        # NOTE: even if SciPy computes the solution "successfully", there might be NaNs
+        # NOTE: in the result, so the test has to check for that as well
+        if np.any(np.isnan(x_ref)):
+            x_ref = None
+
+    except np.linalg.LinAlgError:
+        pass
+
+    # the banded matrix is LU decomposed with the respective Chemotools function
+    lu_fact = lu_banded(
+        l_and_u=l_and_u,
+        ab=ab,
[…]

[…]
