### 2024-05-10 22476c3
tests: wrote separate doctested utility function for banded log-determinant; covered more complete range for banded linalg tests
