### 2024-05-10 d208e0d
fix: removed now wrong Whittaker implementation details
