+++ chemotools/utils/WHITTAKER_IMPLEMENTATION_DETAILS.md
@@ -1,200 +1,0 @@
-# 🧮 Implementation of the Whittaker Henderson smoothing algorithm
-
-The internal linear system to be solved for the Whittaker Henderson smoothing algorithm
-is given by the following set of equations:<br>
-
-$\mathbf{z}=(\mathbf{W}+\tau\cdot\mathbf{D}_{m}^{T}\mathbf{D}_{m})^{-1}\mathbf{W}\mathbf{y}$
-
-where $\mathbf{z}$ is the vector of smoothed values, $\mathbf{W}$ is the diagonal matrix
-of weights, $\tau$ is the smoothing parameter, $\mathbf{D}_{m}$ is the ${m}$-th order
-forward finite difference matrix, and $\mathbf{y}$ is the vector of observed values.
-The matrix $P=\mathbf{D}_{m}^{T}\mathbf{D}_{m}$ is often referred to as the
-*penalty matrix*.<br><br>
-The matrix to invert is symmetric, sparse, banded with ${2\cdot m + 1}$ non-zero
-diagonals, and positive definite, i.e., all its eigenvalues are strictly positive
-($>0$). From an algorithmic point of view, this means that the inversion can be performed
-in ${\mathcal{O}\left(n\right)}$ time, where ${n}$ is the number of observations, by
-using a banded Cholesky decomposition:<br>
-
-$\mathbf{W}+\tau\cdot\mathbf{P}=\mathbf{L}\mathbf{L}^T$
-
-where $\mathbf{L}$ is a lower triangular matrix which shares the same sparsity pattern
-as $\mathbf{W}+\tau\cdot\mathbf{P}$. Inversion of a lower triangular matrix is trivial
-when backward and forward substitution are used.
-
-## ⚠️ Problem
-
-However, all this is only true from a mathematical point of view. In practice,
-floating point arithmetics introduce numerical errors which can lead to an indefinite
-matrix. In this case, the Cholesky decomposition fails and the algorithm cannot be used.
-This happens for relatively small $n$ already when $m$ exceeds 4, but in spectroscopy
-$n > 1000$ is not uncommon and $m = 6$ has been shown helpful in deriving an additional
-set of weights fro $\tau$ to make the smooth spatially adaptive.<br>
-Besides, there is another problem. The penalty matrix
-$\mathbf{P}$ alone is only positive semi-definite since it possesses $m$ zero
-eigenvalues by design. From a mathematical perspective, this would not pose a problem
-since $\mathbf{W}+\tau\cdot\mathbf{P}$ would still be positive definite as $\tau$ tends
-to $+\infty$. Numerically, this is by far not the case because $\tau$-values that are an
-order of $10^{16}$ greater than the order of the weights are already sufficient to make
-the matrix positive semi-definite or even indefinite if some of the small eigenvalues go
-negative in the calculations (for 64-bit float precision).<br>
-On the other hand, as $\tau$ tends to $0$, the matrix can also become ill-conditioned
-as well if some weights are numerically zero due to zero division.<br><br>
-All in all, the banded Cholesky decomposition is not a robust algorithm for solving the
-system of linear equations and even pivoted LU-decomposition suffers from the same
[…]
