### 2024-05-10 428948c
fix: made error message compatible to the questionable Sklearn standard error message
