### 2024-05-10 e3dbf2f
fix: fixed original signal overwrite
