### 2024-05-10 a72900a
refactor: updated Whittaker-like baseline corrections with new Whittaker logic
