### 2024-05-10 5e41a03
refactor: adapted to new internal whittaker implementation
