+++ chemotools/smooth/_whittaker_smooth.py
@@ -111,7 +111,7 @@ class WhittakerSmooth(
         self.n_features_in_ = X.shape[1]
         # ... and all the required attributes for fitting
         self._setup_for_fit(
-            series_size=self.n_features_in_,
+            n_data=self.n_features_in_,
             lam=self.lam,
             differences=self.differences,
         )
