### 2024-05-10 d4a2f1d
refactor: updated Whittaker functionality tests; reverted internal copying
