+++ tests/test_functionality.py
@@ -30,7 +30,7 @@ from chemotools.scatter import (
     StandardNormalVariate,
 )
 from chemotools.smooth import MeanFilter, MedianFilter, WhittakerSmooth
-from chemotools.utils.models import _PENTAPY_AVAILABLE, BandedSolveDecompositions
+from chemotools.utils.models import _PENTAPY_AVAILABLE, BandedSolvers
 from tests.fixtures import reference_airpls  # noqa: F401
 from tests.fixtures import reference_arpls  # noqa: F401
 from tests.fixtures import reference_msc_mean  # noqa: F401
@@ -817,12 +817,17 @@ def test_whittaker_smooth(
     else:
         weights = None
 
+    spectrum_to_fit_original = np.tile(spectrum, reps=reps)
+    spectrum_to_fit = spectrum_to_fit_original.copy()
+
     # Act
     spectrum_corrected = whittaker_smooth.fit_transform(
-        X=np.tile(spectrum, reps=reps), sample_weight=weights
+        X=spectrum_to_fit, sample_weight=weights
     )
 
     # Assert
+    # NOTE: the following test makes sure nothing was overwritten
+    assert np.array_equal(spectrum_to_fit, spectrum_to_fit_original)
     assert np.allclose(
         spectrum_corrected, np.tile(reference_whittaker, reps=reps), atol=1e-8
     )
@@ -836,65 +841,64 @@ def test_whittaker_with_pentapy(
 ):
     # this test is skipped with a warning if pentapy is not installed
     if not _PENTAPY_AVAILABLE:
-        logging.warning("pentapy is not installed")
-        pytest.skip("pentapy is not installed, test cannot be performed")
+        pytest.skip("Pentapy is not installed, test cannot be performed")
     # else nothing
 
     # Arrange
     np.random.seed(42)
     spectrum = np.random.rand(n_samples, 1000)
-    whittaker_smooth = WhittakerSmooth(differences=2)
+    whittaker_smooth = WhittakerSmooth(lam=100.0, differences=2)
+
+    weights = None
     if with_weights and not same_weights_for_all:
         weights = np.ones(shape=(n_samples, len(spectrum[0])))
     elif with_weights and same_weights_for_all:
         weights = np.ones(shape=(len(spectrum[0]),))
-    else:
-        weights = None
 
     # Act with pentapy
     spectrum_corr_pentapy = whittaker_smooth.fit_transform(
-        spectrum, sample_weight=weights
+        X=spectrum, sample_weight=weights
     )
 
     # Assert with pentapy
-    assert (
-        whittaker_smooth._solve(
-            bw=spectrum.transpose(),
-            log_lam=np.log(whittaker_smooth.lam),
-            w=None,
-            mod_squ_fin_diff_mat_lub=whittaker_smooth.base_squ_fw_fin_diff_mat_lub_,
-        )[2]
-        == BandedSolveDecompositions.PENTAPY
-    )
+    # NOTE: the weight is not correct since the test only checks the method
+    solve_method = whittaker_smooth._solve(
+        lam=whittaker_smooth._lam_inter_.fixed_lambda,
+        b_weighted=spectrum.transpose(),
+        w=1.0,
+    )[1]
+    assert solve_method == BandedSolvers.PENTAPY
 
     # Act without pentapy
     whittaker_smooth._WhittakerLikeSolver__allow_pentapy = False  # type: ignore
-    spectrum_corr_scipy = whittaker_smooth.fit_transform(
+    spectrum_corr_factorized_solve = whittaker_smooth.fit_transform(
         spectrum, sample_weight=weights
[…]
