### 2024-05-10 664e66e
refactor: moved `pentapy` import check into dedicated `_runtime`-module
