+++ tests/test_functionality.py
@@ -1,9 +1,8 @@
-import logging
-
 import numpy as np
 import pandas as pd
 import pytest
 
+from chemotools._runtime import PENTAPY_AVAILABLE
 from chemotools.augmentation import (
     BaselineShift,
     ExponentialNoise,
@@ -30,7 +29,7 @@ from chemotools.scatter import (
     StandardNormalVariate,
 )
 from chemotools.smooth import MeanFilter, MedianFilter, WhittakerSmooth
-from chemotools.utils.models import _PENTAPY_AVAILABLE, BandedSolvers
+from chemotools.utils.models import BandedSolvers
 from tests.fixtures import reference_airpls  # noqa: F401
 from tests.fixtures import reference_arpls  # noqa: F401
 from tests.fixtures import reference_msc_mean  # noqa: F401
@@ -840,7 +839,7 @@ def test_whittaker_with_pentapy(
     n_samples: int, with_weights: bool, same_weights_for_all: bool
 ):
     # this test is skipped with a warning if pentapy is not installed
-    if not _PENTAPY_AVAILABLE:
+    if not PENTAPY_AVAILABLE:
         pytest.skip("Pentapy is not installed, test cannot be performed")
     # else nothing
 
