### 2024-05-10 67694d9
refactor: raise error when weights are not provided for log marginal likelihood
