### 2024-05-10 200e45e
fix: fixed type hints breaking Python 3.9 compatibility
