### 2024-05-10 fc206a1
refactor/fix: removed dead functions; excluded hard-to-produce-edge-cases from coverage; removed unused imports
