+++ chemotools/utils/banded_linalg.py
@@ -14,14 +14,9 @@ factorizations.
 
 ### Imports ###
 
-from numbers import Integral
-from typing import Union
-
 import numpy as np
 from numpy.typing import ArrayLike
 from scipy.linalg import lapack
-from scipy.sparse import spmatrix
-from sklearn.utils import check_array, check_scalar
 
 from chemotools.utils.models import BandedLUFactorization
 
@@ -49,52 +44,6 @@ def _datacopied(arr, original):
         return False
 
     return arr.base is None
-
-
-def _check_full_arr_n_diag_counts_for_lu_banded(
-    a_shape: tuple[int, int],
-    l_and_u: tuple[int, int],
-) -> None:
-    """Validates the shape of a full array and the number of sub- and superdiagonals
-    for LU decomposition of a banded (sparse) matrix.
-    """
-    num_rows, num_cols = a_shape
-    num_low_diags, num_upp_diags = l_and_u
-
-    check_scalar(
-        x=num_rows,
-        name="num_rows",
-        target_type=Integral,
-        min_val=1,
-        include_boundaries="left",
-    )
-    check_scalar(
-        x=num_cols,
-        name="num_cols",
-        target_type=Integral,
-        min_val=1,
-        include_boundaries="left",
-    )
-    check_scalar(
-        x=num_low_diags,
-        name="num_low_diags",
-        target_type=Integral,
-        min_val=0,
-        max_val=num_rows - 1,
-        include_boundaries="both",
-    )
-    check_scalar(
-        x=num_upp_diags,
-        name="num_upp_diags",
-        target_type=Integral,
-        min_val=0,
-        max_val=num_rows - 1,
-        include_boundaries="both",
-    )
-
-    if num_rows != num_cols:
-        raise ValueError(f"\nThe matrix must be square, but it has shape {a_shape}.")
-
 
 
 def conv_upper_chol_banded_to_lu_banded_storage(
@@ -184,103 +133,6 @@ def conv_upper_chol_banded_to_lu_banded_storage(
     return l_and_u, np.row_stack((ab, ab_subdiags))
 
 
-def conv_to_lu_banded_storage(
-    a: Union[np.ndarray, spmatrix],
-    l_and_u: tuple[int, int],
-) -> np.ndarray:
-    """Converts a (sparse) square banded matrix A to its banded storage required for
-    LU decomposition in LAPACK-routines like the function ``lu_banded`` or SciPy's
-    ``solve_banded``. This format is identical for pentapy where it is referred to as
-    "column-wise flattened".
-    Cholesky-decompositions require a different format.
-
-    Parameters
-    ----------
-    a : np.ndarray or sparse matrix of shape (n, n)
-        A square banded NumPy-2D-Array or SciPy sparse matrix. "Square" means that the
-        row count equals the column count while "banded" implies that only the main
-        diagonal and a few sub- and/or superdiagonals are non-zero (see `l_and_u`).
-    l_and_u : tuple[int, int]
-        The number of "non-zero" sub- (first) and superdiagonals (second element) aside
-        the main diagonal which does not need to be considered here. "Non-zero" can be
-        a bit misleading in this context. These numbers should count up to the diagonal
-        after which all following diagonals are zero. Zero-diagonals that come before
-        still need to be included.
-        Wrong specification of this can lead to non-zero-diagonals being ignored or
-        zero-diagonals being included which corrupts the results or reduces the
-        performance.
-
-    Returns
-    -------
-    ab : np.ndarray of shape (l_and_u[0] + 1 + l_and_u[1], n)
[…]
