+++ pyproject.toml
@@ -5,6 +5,11 @@ requires = [
 ]
 build-backend = "setuptools.build_meta"
 
+[tool.coverage.run]
+omit = [
+    "chemotools/_runtime/*",
+    ]
+
 [tool.ruff]
 # Enable pycodestyle (`E`), Pyflakes (`F`) checks.
 select = ["E", "F"]
