### 2024-05-10 a8ecb7d
tests: added parametrized parallelizable tests for the lambda-value of the Whittaker smoother
