+++ tests/test_for_utils/test_models.py
@@ -1,0 +1,252 @@
+"""
+Test suite for the utility models in the :mod:`chemotools.utils.models` module.
+
+"""
+
+### Imports ###
+
+from typing import List, Tuple, Union
+
+import numpy as np
+import pytest
+
+from chemotools.utils.models import WhittakerSmoothLambda, WhittakerSmoothMethods
+from tests.test_for_utils.utils import ExpectedWhittakerSmoothLambda, float_is_bit_equal
+
+### Type aliases ###
+
+_Numeric = Union[float, int]
+_LambdaValueNumeric = Union[_Numeric, Tuple[_Numeric, _Numeric]]
+_LambdaValueNumericOrFlawed = Union[_LambdaValueNumeric, str]
+_WhittakerMethod = Union[str, WhittakerSmoothMethods]
+_WhittakerMethodSequence = List[_WhittakerMethod]
+_ExpectedLambdaResult = Union[
+    ExpectedWhittakerSmoothLambda,
+    ValueError,
+    TypeError,
+]
+_LambdaTestCombination = Tuple[
+    _LambdaValueNumericOrFlawed,
+    _WhittakerMethodSequence,
+    _ExpectedLambdaResult,
+]
+
+### Global constants ###
+
+_NAN: float = float("nan")
+_FIXED_WHITTAKER_METHODS: _WhittakerMethodSequence = [
+    "fixed",
+    WhittakerSmoothMethods.FIXED,
+]
+_LOGML_WHITTAKER_METHODS: _WhittakerMethodSequence = [
+    "logml",
+    WhittakerSmoothMethods.LOGML,
+]
+# NOTE: "aauto" is not a typo, but helps to not confuse it with "all"
+_aauto_whittaker_methods: _WhittakerMethodSequence = _LOGML_WHITTAKER_METHODS + []
+_all_whittaker_methods: _WhittakerMethodSequence = (
+    _FIXED_WHITTAKER_METHODS + _aauto_whittaker_methods
+)
+
+
+### Test Suite ###
+
+
+@pytest.mark.parametrize(
+    "combination",
+    [
+        (  # Number 0
+            100.0,
+            _FIXED_WHITTAKER_METHODS,
+            ExpectedWhittakerSmoothLambda(
+                fixed_lambda=100.0,
+                auto_bounds=(_NAN, _NAN),
+                fit_auto=False,
+                method_used=WhittakerSmoothMethods.FIXED,
+                log_auto_bounds=(_NAN, _NAN),
+            ),
+        ),
+        (  # Number 1
+            (100.0, 100.0),
+            _FIXED_WHITTAKER_METHODS,
+            ExpectedWhittakerSmoothLambda(
+                fixed_lambda=100.0,
+                auto_bounds=(_NAN, _NAN),
+                fit_auto=False,
+                method_used=WhittakerSmoothMethods.FIXED,
+                log_auto_bounds=(_NAN, _NAN),
+            ),
+        ),
+        (  # Number 2
+            (100.0, 100.000001),
+            _FIXED_WHITTAKER_METHODS,
+            ExpectedWhittakerSmoothLambda(
+                fixed_lambda=100.000001,
+                auto_bounds=(_NAN, _NAN),
+                fit_auto=False,
+                method_used=WhittakerSmoothMethods.FIXED,
+                log_auto_bounds=(_NAN, _NAN),
+            ),
+        ),
+        (  # Number 3
+            (100.0, 100.000001),
+            _aauto_whittaker_methods,
+            ExpectedWhittakerSmoothLambda(
+                fixed_lambda=100.000001,
+                auto_bounds=(_NAN, _NAN),
+                fit_auto=False,
[…]
