### 2024-05-10 365c5dd
tests: added utility functions for model testing
