### 2024-05-11 9140e92
style: made class heritage more structured
