### 2024-05-11 199eac4
refactor: split up whittaker base to make it more modular, flexible, future-suited, and readable - part I basic solvers
