+++ chemotools/utils/whittaker_base/solvers.py
@@ -1,0 +1,216 @@
+"""
+This submodule contains the solver functions used by the ``WhittakerLikeSolver`` class
+that would have cluttered the class implementation.
+
+"""
+
+### Imports ###
+
+
+from typing import Union
+
+import numpy as np
+
+from chemotools._runtime import PENTAPY_AVAILABLE
+from chemotools.utils import banded_linalg as bla
+from chemotools.utils import models
+
+if PENTAPY_AVAILABLE:
+    import pentapy as pp
+
+### Type Aliases ###
+
+_Factorization = Union[models.BandedLUFactorization, models.BandedPentapyFactorization]
+
+### Functions ###
+
+
+def solve_pentapy(a_banded: np.ndarray, b_weighted: np.ndarray) -> np.ndarray:
+    """
+    Solves the linear system of equations ``(W + lam * D.T @ D) @ x = W @ b`` with the
+    ``pentapy`` package. This is the same as solving the linear system ``A @ x = b``
+    where ``A = W + lam * D.T @ D`` and ``b = W @ b``.
+
+    Notes
+    -----
+    Pentapy does not (maybe yet) allow for 2D right-hand side matrices, so the
+    solution is computed for each column of ``bw`` separately.
+
+    """
+
+    # for 1-dimensional right-hand side vectors, the solution is computed directly
+    if b_weighted.ndim == 1:
+        return pp.solve(
+            mat=a_banded,
+            rhs=b_weighted,
+            is_flat=True,
+            index_row_wise=False,
+            solver=1,
+        )
+
+    # for 2-dimensional right-hand side matrices, the solution is computed for each
+    # column separately
+    else:
+        # NOTE: the solutions are first written into the rows of the solution matrix
+        #       because row-access is more efficient for C-contiguous arrays;
+        #       afterwards, the solution matrix is transposed
+        solution = np.empty(shape=(b_weighted.shape[1], b_weighted.shape[0]))
+        for iter_j in range(0, b_weighted.shape[1]):
+            solution[iter_j, ::] = pp.solve(
+                mat=a_banded,
+                rhs=b_weighted[::, iter_j],
+                is_flat=True,
+                index_row_wise=False,
+                solver=1,
+            )
+
+        return solution.transpose()
+
+
+def solve_ppivoted_lu(
+    l_and_u: bla.LAndUBandCounts,
+    a_banded: np.ndarray,
+    b_weighted: np.ndarray,
+) -> tuple[np.ndarray, models.BandedLUFactorization]:
+    """
+    Solves the linear system of equations ``(W + lam * D.T @ D) @ x = W @ b`` with a
+    partially pivoted LU decomposition. This is the same as solving the linear system
+    ``A @ x = b`` where ``A = W + lam * D.T @ D`` and ``b = W @ b``.
+
+    If the LU decomposition fails, a ``LinAlgError`` is raised which is fatal since
+    the next level of escalation would be using a QR-decomposition which is not
+    implemented (yet).
+
+    """
+
+    lub_factorization = bla.lu_banded(
+        l_and_u=l_and_u,
+        ab=a_banded,
+        check_finite=False,
+    )
+    return (
+        bla.lu_solve_banded(
+            lub_factorization=lub_factorization,
+            b=b_weighted,
+            check_finite=False,
+            overwrite_b=True,
+        ),
+        lub_factorization,
+    )
+
+
+def solve_normal_equations(
+    lam: float,
+    differences: int,
+    l_and_u: bla.LAndUBandCounts,
+    penalty_mat_banded: np.ndarray,
+    b_weighted: np.ndarray,
+    w: Union[float, np.ndarray],
+    pentapy_enabled: bool,
[…]
