### 2024-05-11 a3b5af6
refactor: split up whittaker base to make it more modular, flexible, future-suited, and readable - part II log marginal likelihood
