+++ chemotools/utils/whittaker_base/main.py
@@ -15,17 +15,11 @@ import numpy as np
 
 from chemotools._runtime import PENTAPY_AVAILABLE
 from chemotools.utils import models
-from chemotools.utils.banded_linalg import LAndUBandCounts, slogdet_lu_banded
+from chemotools.utils.banded_linalg import LAndUBandCounts
 from chemotools.utils.whittaker_base import auto_lambda as auto
 from chemotools.utils.whittaker_base import initialisation as init
 from chemotools.utils.whittaker_base import solvers
 from chemotools.utils.whittaker_base.misc import get_weight_generator
-
-### Type Aliases ###
-
-_Factorization = Union[models.BandedLUFactorization, models.BandedPentapyFactorization]
-_FactorizationForLogMarginalLikelihood = models.BandedLUFactorization
-
 
 ### Class Implementation ###
 
@@ -88,8 +82,6 @@ class WhittakerLikeSolver:
 
     """  # noqa: E501
 
-    __LN_TWO_PI: float = 1.8378770664093453
-    __LN_TEN: float = 2.302585092994046
     __dtype: type = np.float64
     __allow_pentapy: bool = True
     __zero_weight_tol: float = 1e-10
@@ -168,7 +160,7 @@ class WhittakerLikeSolver:
         lam: float,
         b_weighted: np.ndarray,
         w: Union[float, np.ndarray],
-    ) -> tuple[np.ndarray, models.BandedSolvers, _Factorization]:
+    ) -> tuple[np.ndarray, models.BandedSolvers, auto._Factorization]:
         """
         Internal wrapper for the solver methods to solve the linear system of equations
         for the Whittaker-like smoother.
@@ -192,133 +184,7 @@ class WhittakerLikeSolver:
             pentapy_enabled=self._pentapy_enabled_,
         )
 
-    ### Auxiliary Methods to prepare the data for the solver ###
-
-    def calc_wrss(
-        self, b: np.ndarray, b_smooth: np.ndarray, w: Union[float, np.ndarray]
-    ) -> float:
-        """
-        Computes the (weighted) Sum of Squared Residuals (w)RSS between the original and
-        the smoothed series.
-
-        """
-
-        # Case 1: no weights are provided
-        if isinstance(w, float):
-            return np.square(b - b_smooth).sum()
-
-        # Case 2: weights are provided
-        return (w * np.square(b - b_smooth)).sum()
-
-    def _calc_log_marginal_likelihood(
-        self,
-        factorization: _FactorizationForLogMarginalLikelihood,
-        log_lam: float,
-        lam: float,
-        b: np.ndarray,
-        b_smooth: np.ndarray,
-        w: Union[float, np.ndarray],
-        w_plus_penalty_plus_n_samples_term: float,
-    ) -> float:
-        """
-        Computes the log marginal likelihood for the automatic fitting of the penalty
-        weight lambda. For the definitions used (and manipulated here), please refer to
-        the Notes section.
-
-        Parameters
-        ----------
-        factorization : BandedLUFactorization
-            The factorization of the matrix to solve the linear system of equations,
-            i.e., ``W + lambda * D.T @ D`` from the description above.
-            Currently, only partially pivoted banded LU decompositions can be used to
-            compute the log marginal likelihood.
-        log_lam : float
-            The natural logarithm of the penalty weight lambda used for the smoothing.
-        lam : float
-            The penalty weight lambda used for the smoothing, i.e., ``exp(log_lam)``.
-        b, b_smooth : ndarray of shape (m,)
-            The original series and its smoothed counterpart.
-        w : float or ndarray of shape (m,)
-            The weights to use for the smoothing.
-        w_plus_penalty_plus_n_samples_term : float
-            The last term of the log marginal likelihood that is constant since it
[…]
