+++ chemotools/utils/whittaker_base/auto_lambda/logml.py
@@ -1,0 +1,191 @@
+"""
+This submodule contains the functions used for the automated fitting of the penalty
+weight lambda within the ``WhittakerLikeSolver`` class based on the log marginal
+likelihood that would have cluttered the class implementation.
+
+"""
+
+### Imports ###
+
+from typing import Union
+
+import numpy as np
+
+from chemotools.utils import banded_linalg as bla
+from chemotools.utils import models
+from chemotools.utils.whittaker_base.auto_lambda.shared import get_smooth_wrss
+
+### Constants ###
+
+_LN_TWO_PI: float = 1.8378770664093453  # ln(2 * pi)
+
+### Type Aliases ###
+
+# TODO: add QR factorization
+_FactorizationForLogMarginalLikelihood = models.BandedLUFactorization
+
+### Functions ###
+
+
+def get_log_marginal_likelihood_constant_term(
+    differences: int,
+    penalty_mat_log_pseudo_det: float,
+    w: np.ndarray,
+    zero_weight_tol: float,
+) -> float:
+    """
+    Computes the constant term of the log marginal likelihood for the automatic fitting
+    of the penalty weight lambda, i.e.,
+
+    ``(n^ - d) * ln(2 * pi) - ln(pseudo_det(W)) - ln(pseudo_det(D.T @ D))``
+
+    or better
+
+    ``(n^ - d) * ln(2 * pi) - ln(pseudo_det(W)) - ln(det(D @ D.T))``
+
+    with:
+
+    - ``ln`` as the natural logarithm,
+    - ``pseudo_det(A)`` as the pseudo-determinant of the matrix ``A``, i.e., the product
+        of its non-zero eigenvalues,
+    - ``det(A)`` as the determinant of the matrix ``A``, i.e., the product of its
+        eigenvalues,
+    - ``W`` as the diagonal matrix with the weights on the main diagonal,
+    - ``D.T @ D`` as the squared forward finite differences matrix,
+    - ``D @ D.T`` as the flipped squared forward finite differences matrix,
+    - ``d`` as the difference order used for the smoothing, and
+    - ``n^`` as the number of data points with non-zero weights in the series to smooth.
+
+    It should be noted that ``pseudo_det(D.T @ D)`` is replaced by ``det(D @ D.T)`` here
+    because the latter is not rank-deficient.
+
+    """
+
+    # first, the constant terms of the log marginal likelihood are computed starting
+    # from the log pseudo-determinant of the weight matrix, i.e., the product of the
+    # non-zero elements of the weight vector
+    nonzero_w_idxs = np.where(w > w.max() * zero_weight_tol)[0]
+    nnz_w = nonzero_w_idxs.size
+    log_pseudo_det_w = np.log(w[nonzero_w_idxs]).sum()
+
+    # the constant term of the log marginal likelihood is computed
+    return (
+        (nnz_w - differences) * _LN_TWO_PI
+        - log_pseudo_det_w
+        - penalty_mat_log_pseudo_det
+    )
+
+
+def get_log_marginal_likelihood(
+    factorization: _FactorizationForLogMarginalLikelihood,
+    log_lam: float,
+    lam: float,
+    differences: int,
+    diff_kernel_flipped: np.ndarray,
+    b: np.ndarray,
+    b_smooth: np.ndarray,
+    w: Union[float, np.ndarray],
+    w_plus_penalty_plus_n_samples_term: float,
+) -> float:
+    """
+    Computes the log marginal likelihood for the automatic fitting of the penalty
+    weight lambda. For the definitions used (and manipulated here), please refer to
+    the Notes section.
+
+    Parameters
+    ----------
+    factorization : BandedLUFactorization
+        The factorization of the matrix to solve the linear system of equations,
+        i.e., ``W + lambda * D.T @ D`` from the description above.
[…]
