+++ chemotools/utils/whittaker_base/auto_lambda/__init__.py
@@ -7,6 +7,13 @@ class implementation.
 
 ### Imports ###
 
-from chemotools.utils.whittaker_base.auto_lambda.optimization import (
+from chemotools.utils.whittaker_base.auto_lambda.logml import (  # noqa: F401
+    get_log_marginal_likelihood,
+    get_log_marginal_likelihood_constant_term,
+)
+from chemotools.utils.whittaker_base.auto_lambda.optimization import (  # noqa: F401
     get_optimized_lambda,
-)  # noqa: F401
+)
+from chemotools.utils.whittaker_base.auto_lambda.shared import (  # noqa: F401
+    _Factorization,
+)
