### 2024-05-11 1c709d3
tests/refactor: simplified model tests
