+++ tests/test_for_utils/utils.py
@@ -19,25 +19,7 @@ from scipy.linalg import eigvals_banded
 from scipy.sparse import csr_matrix
 from scipy.sparse import diags as sp_diags
 
-from chemotools.utils.models import WhittakerSmoothMethods
-
-### Dataclasses ###
-
-
-@dataclass
-class ExpectedWhittakerSmoothLambda:
-    """
-    Dataclass for checking the expected results for the class :class:`WhittakerSmoothLambda`
-    from the module :mod:`chemotools.utils.models`.
-
-    """  # noqa: E501
-
-    fixed_lambda: float
-    auto_bounds: Tuple[float, float]
-    fit_auto: bool
-    method_used: WhittakerSmoothMethods
-    log_auto_bounds: Tuple[float, float] = (0.0, 0.0)
-
+from chemotools.utils import models
 
 ### Utility Functions ###
 
@@ -771,6 +753,54 @@ def get_banded_slogdet(ab: np.ndarray) -> Tuple[float, float]:
         logabsdet = np.log(np.abs(eigvals)).sum()  # type: ignore
 
     return sign, logabsdet
+
+
+### Dataclasses ###
+
+
+@dataclass
+class ExpectedWhittakerSmoothLambda:
+    """
+    Dataclass for checking the expected results for the class :class:`WhittakerSmoothLambda`
+    from the module :mod:`chemotools.utils.models`.
+
+    """  # noqa: E501
+
+    fixed_lambda: float
+    auto_bounds: Tuple[float, float]
+    fit_auto: bool
+    method_used: models.WhittakerSmoothMethods
+    log_auto_bounds: Tuple[float, float] = (0.0, 0.0)
+
+    def assert_is_equal_to(self, other: models.WhittakerSmoothLambda) -> None:
+        """
+        Checks if the current instance is equal to another instance of the same class.
+
+        """
+
+        assert other.fit_auto is self.fit_auto
+        assert other.method_used == self.method_used
+        # NOTE: since NAN-values are used, the comparison is split into two parts for
+        #       the fixed lambda value and each of the bounds
+        assert float_is_bit_equal(
+            value=other.fixed_lambda,
+            reference=self.fixed_lambda,
+        )
+        assert float_is_bit_equal(
+            value=other.auto_bounds[0], reference=self.auto_bounds[0]
+        )
+        assert float_is_bit_equal(
+            value=other.auto_bounds[1],
+            reference=self.auto_bounds[1],
+        )
+        assert float_is_bit_equal(
+            value=other.log_auto_bounds[0],
+            reference=self.log_auto_bounds[0],
+        )
+        assert float_is_bit_equal(
+            value=other.log_auto_bounds[1],
+            reference=self.log_auto_bounds[1],
+        )
 
 
 ### Doctests ###
