+++ tests/test_for_utils/test_models.py
@@ -5,25 +5,25 @@ Test suite for the utility models in the :mod:`chemotools.utils.models` module.
 
 ### Imports ###
 
-from typing import List, Tuple, Union
+from typing import List, Tuple, Type, Union
 
 import numpy as np
 import pytest
 
-from chemotools.utils.models import WhittakerSmoothLambda, WhittakerSmoothMethods
-from tests.test_for_utils.utils import ExpectedWhittakerSmoothLambda, float_is_bit_equal
+from chemotools.utils import models
+from tests.test_for_utils.utils import ExpectedWhittakerSmoothLambda
 
 ### Type aliases ###
 
 _Numeric = Union[float, int]
 _LambdaValueNumeric = Union[_Numeric, Tuple[_Numeric, _Numeric]]
 _LambdaValueNumericOrFlawed = Union[_LambdaValueNumeric, str]
-_WhittakerMethod = Union[str, WhittakerSmoothMethods]
+_WhittakerMethod = Union[str, models.WhittakerSmoothMethods]
 _WhittakerMethodSequence = List[_WhittakerMethod]
 _ExpectedLambdaResult = Union[
     ExpectedWhittakerSmoothLambda,
-    ValueError,
-    TypeError,
+    Type[ValueError],
+    Type[TypeError],
 ]
 _LambdaTestCombination = Tuple[
     _LambdaValueNumericOrFlawed,
@@ -31,16 +31,16 @@ _LambdaTestCombination = Tuple[
     _ExpectedLambdaResult,
 ]
 
-### Global constants ###
+### Constants ###
 
 _NAN: float = float("nan")
 _FIXED_WHITTAKER_METHODS: _WhittakerMethodSequence = [
     "fixed",
-    WhittakerSmoothMethods.FIXED,
+    models.WhittakerSmoothMethods.FIXED,
 ]
 _LOGML_WHITTAKER_METHODS: _WhittakerMethodSequence = [
     "logml",
-    WhittakerSmoothMethods.LOGML,
+    models.WhittakerSmoothMethods.LOGML,
 ]
 # NOTE: "aauto" is not a typo, but helps to not confuse it with "all"
 _aauto_whittaker_methods: _WhittakerMethodSequence = _LOGML_WHITTAKER_METHODS + []
@@ -62,7 +62,7 @@ _all_whittaker_methods: _WhittakerMethodSequence = (
                 fixed_lambda=100.0,
                 auto_bounds=(_NAN, _NAN),
                 fit_auto=False,
-                method_used=WhittakerSmoothMethods.FIXED,
+                method_used=models.WhittakerSmoothMethods.FIXED,
                 log_auto_bounds=(_NAN, _NAN),
             ),
         ),
@@ -73,7 +73,7 @@ _all_whittaker_methods: _WhittakerMethodSequence = (
                 fixed_lambda=100.0,
                 auto_bounds=(_NAN, _NAN),
                 fit_auto=False,
-                method_used=WhittakerSmoothMethods.FIXED,
+                method_used=models.WhittakerSmoothMethods.FIXED,
                 log_auto_bounds=(_NAN, _NAN),
             ),
         ),
@@ -84,7 +84,7 @@ _all_whittaker_methods: _WhittakerMethodSequence = (
                 fixed_lambda=100.000001,
                 auto_bounds=(_NAN, _NAN),
                 fit_auto=False,
-                method_used=WhittakerSmoothMethods.FIXED,
+                method_used=models.WhittakerSmoothMethods.FIXED,
                 log_auto_bounds=(_NAN, _NAN),
             ),
         ),
@@ -95,7 +95,7 @@ _all_whittaker_methods: _WhittakerMethodSequence = (
                 fixed_lambda=100.000001,
[…]
