### 2024-05-11 63b1217
tests: added clearer comments to model tests
