+++ tests/test_for_utils/test_models.py
@@ -20,11 +20,7 @@ _LambdaValueNumeric = Union[_Numeric, Tuple[_Numeric, _Numeric]]
 _LambdaValueNumericOrFlawed = Union[_LambdaValueNumeric, str]
 _WhittakerMethod = Union[str, models.WhittakerSmoothMethods]
 _WhittakerMethodSequence = List[_WhittakerMethod]
-_ExpectedLambdaResult = Union[
-    ExpectedWhittakerSmoothLambda,
-    Type[ValueError],
-    Type[TypeError],
-]
+_ExpectedLambdaResult = Union[ExpectedWhittakerSmoothLambda, Type[Exception]]
 _LambdaTestCombination = Tuple[
     _LambdaValueNumericOrFlawed,
     _WhittakerMethodSequence,
@@ -55,7 +51,7 @@ _all_whittaker_methods: _WhittakerMethodSequence = (
 @pytest.mark.parametrize(
     "combination",
     [
-        (  # Number 0
+        (  # Number 0 (fixed float; fixed method)
             100.0,
             _FIXED_WHITTAKER_METHODS,
             ExpectedWhittakerSmoothLambda(
@@ -66,7 +62,18 @@ _all_whittaker_methods: _WhittakerMethodSequence = (
                 log_auto_bounds=(_NAN, _NAN),
             ),
         ),
-        (  # Number 1
+        (  # Number 1 (fixed integer; fixed method)
+            100,
+            _FIXED_WHITTAKER_METHODS,
+            ExpectedWhittakerSmoothLambda(
+                fixed_lambda=100.0,
+                auto_bounds=(_NAN, _NAN),
+                fit_auto=False,
+                method_used=models.WhittakerSmoothMethods.FIXED,
+                log_auto_bounds=(_NAN, _NAN),
+            ),
+        ),
+        (  # Number 2 (coinciding floats; fixed method)
             (100.0, 100.0),
             _FIXED_WHITTAKER_METHODS,
             ExpectedWhittakerSmoothLambda(
@@ -77,7 +84,18 @@ _all_whittaker_methods: _WhittakerMethodSequence = (
                 log_auto_bounds=(_NAN, _NAN),
             ),
         ),
-        (  # Number 2
+        (  # Number 3 (coinciding integers; fixed method)
+            (100, 100),
+            _FIXED_WHITTAKER_METHODS,
+            ExpectedWhittakerSmoothLambda(
+                fixed_lambda=100.0,
+                auto_bounds=(_NAN, _NAN),
+                fit_auto=False,
+                method_used=models.WhittakerSmoothMethods.FIXED,
+                log_auto_bounds=(_NAN, _NAN),
+            ),
+        ),
+        (  # Number 4 (virtually coinciding floats; fixed method)
             (100.0, 100.000001),
             _FIXED_WHITTAKER_METHODS,
             ExpectedWhittakerSmoothLambda(
@@ -88,7 +106,7 @@ _all_whittaker_methods: _WhittakerMethodSequence = (
                 log_auto_bounds=(_NAN, _NAN),
             ),
         ),
-        (  # Number 3
+        (  # Number 5 (virtually coinciding floats; automated methods)
             (100.0, 100.000001),
             _aauto_whittaker_methods,
             ExpectedWhittakerSmoothLambda(
@@ -99,7 +117,7 @@ _all_whittaker_methods: _WhittakerMethodSequence = (
                 log_auto_bounds=(_NAN, _NAN),
             ),
         ),
-        (  # Number 4
+        (  # Number 6 (flipped virtually coinciding floats; fixed method)
             (100.000001, 100.0),
             _FIXED_WHITTAKER_METHODS,
[…]
