### 2024-05-11 5ec6487
tests/refactor: removed `numpy` from model tests
