### 2024-05-11 70e9322
tests/refactor: renamed `Numeric` to `RealNumeric`
