### 2024-05-11 3fbbd7a
tests/refactor: slightly reduced type aliases
