### 2024-05-11 e57ab4a
refactor: made weight iterator less error prone and better testable
