### 2024-05-11 fcb957b
tests/feat/refactor: added fixture for automated lambda estimation of whittaker smoother; formatted fixtures; changed type hints
