+++ tests/fixtures.py
@@ -1,4 +1,5 @@
 import os
+from typing import List
 
 import numpy as np
 import pytest
@@ -9,88 +10,119 @@ path_to_resources = os.path.join(test_directory, "resources")
 
 
 @pytest.fixture
-def spectrum() -> list[np.ndarray]:
+def spectrum() -> List[np.ndarray]:
     return [
         np.loadtxt(
-            os.path.join(path_to_resources, "spectrum.csv"), delimiter=","
-        ).tolist()
+            os.path.join(path_to_resources, "spectrum.csv"),
+            delimiter=",",
+        )
     ]
 
 
 @pytest.fixture
-def spectrum_arpls() -> list[np.ndarray]:
+def spectrum_arpls() -> List[np.ndarray]:
     return [
         np.loadtxt(
-            os.path.join(path_to_resources, "spectrum_arpls.csv"), delimiter=","
-        ).tolist()
+            os.path.join(path_to_resources, "spectrum_arpls.csv"),
+            delimiter=",",
+        )
     ]
 
 
 @pytest.fixture
-def reference_airpls() -> list[np.ndarray]:
+def reference_airpls() -> List[np.ndarray]:
     return [
         np.loadtxt(
-            os.path.join(path_to_resources, "reference_airpls.csv"), delimiter=","
-        ).tolist()
+            os.path.join(path_to_resources, "reference_airpls.csv"),
+            delimiter=",",
+        )
     ]
 
 
 @pytest.fixture
-def reference_arpls() -> list[np.ndarray]:
+def reference_arpls() -> List[np.ndarray]:
     return [
         np.loadtxt(
-            os.path.join(path_to_resources, "reference_arpls.csv"), delimiter=","
-        ).tolist()
+            os.path.join(path_to_resources, "reference_arpls.csv"),
+            delimiter=",",
+        )
     ]
 
 
 @pytest.fixture
-def reference_msc_mean() -> list[np.ndarray]:
+def reference_msc_mean() -> List[np.ndarray]:
     return [
         np.loadtxt(
-            os.path.join(path_to_resources, "reference_msc_mean.csv"), delimiter=","
-        ).tolist()
+            os.path.join(path_to_resources, "reference_msc_mean.csv"),
+            delimiter=",",
+        )
     ]
 
 
 @pytest.fixture
-def reference_msc_median() -> list[np.ndarray]:
+def reference_msc_median() -> List[np.ndarray]:
     return [
         np.loadtxt(
-            os.path.join(path_to_resources, "reference_msc_median.csv"), delimiter=","
-        ).tolist()
+            os.path.join(path_to_resources, "reference_msc_median.csv"),
+            delimiter=",",
+        )
     ]
 
 
 @pytest.fixture
-def reference_sg_15_2() -> list[np.ndarray]:
+def reference_sg_15_2() -> List[np.ndarray]:
     return [
         np.loadtxt(
-            os.path.join(path_to_resources, "reference_sg_15_2.csv"), delimiter=","
-        ).tolist()
+            os.path.join(path_to_resources, "reference_sg_15_2.csv"),
+            delimiter=",",
+        )
     ]
 
 
 @pytest.fixture
-def reference_snv() -> list[np.ndarray]:
+def reference_snv() -> List[np.ndarray]:
     return [
         np.loadtxt(
-            os.path.join(path_to_resources, "reference_snv.csv"), delimiter=","
-        ).tolist()
+            os.path.join(path_to_resources, "reference_snv.csv"),
+            delimiter=",",
+        )
     ]
 
 
 @pytest.fixture
-def reference_whittaker() -> list[np.ndarray]:
[…]
