### 2024-05-11 0037c71
fix: fixed fatal repetition of upper Cholesky to LU banded storage conversion in log pseudodet computation
