### 2024-05-11 1841b8e
refactor: issue warning on too high difference order; added more information for warning; fixed deprecated NumPy to float warning
