### 2024-05-11 9bb29fe
fix: removed dead 1D-weight generator branch
