### 2024-05-11 4ed784f
tests/fix: removed dead direct call of fixture that caused an error
