### 2024-05-11 c48445c
refactor: made warning an explicit `UserWarning`
