### 2024-05-11 16ef45c
fix: fixed internally broken wrong use of unchecked weights in weight checker; made error messages more descriptive
