### 2024-05-12 3d54add
doc: added TODO for why if statement was excluded from coverage
