### 2024-05-12 a5a5864
tests: added input checker tests
