+++ tests/test_for_utils/test_check_inputs.py
@@ -1,0 +1,204 @@
+"""
+Test suite for the utility models in the :mod:`chemotools.utils.check_inputs` module.
+
+"""
+
+### Imports ###
+
+from typing import Optional, Tuple, Type, Union
+
+import numpy as np
+import pytest
+
+from chemotools.utils.check_inputs import check_weights
+
+### Test Suite ###
+
+
+@pytest.mark.parametrize(
+    "combination",
+    [
+        (  # Number 0 (no weights; for all)
+            None,
+            (None, True),
+        ),
+        (  # Number 1 (valid 1D-weights; for all)
+            np.array([1.0, 2.0, 3.0]),
+            (np.array([[1.0, 2.0, 3.0]]), True),
+        ),
+        (  # Number 2 (valid 2D-weights; for all)
+            np.array([[1.0, 2.0, 3.0]]),
+            (np.array([[1.0, 2.0, 3.0]]), True),
+        ),
+        (  # Number 3 (valid 2D-weights; individual)
+            np.array([[1.0, 2.0, 3.0], [4.0, 5.0, 6.0], [7.0, 8.0, 9.0]]),
+            (np.array([[1.0, 2.0, 3.0], [4.0, 5.0, 6.0], [7.0, 8.0, 9.0]]), False),
+        ),
+        (  # Number 4 (invalid 1D-weights with wrong column number; for all)
+            np.array([1.0, 2.0, 3.0, 4.0]),
+            ValueError,
+        ),
+        (  # Number 5 (invalid 1D-weights with wrong column number; for all)
+            np.array([1.0, 2.0]),
+            ValueError,
+        ),
+        (  # Number 6 (invalid 2D-weights with wrong column number; for all)
+            np.array([[1.0, 2.0, 3.0, 4.0]]),
+            ValueError,
+        ),
+        (  # Number 7 (invalid 2D-weights with wrong column number; for all)
+            np.array([[1.0, 2.0]]),
+            ValueError,
+        ),
+        (  # Number 8 (invalid 2D-weights with wrong row number; individual)
+            np.array([[1.0, 2.0, 3.0], [4.0, 5.0, 6.0]]),
+            ValueError,
+        ),
+        (  # Number 9 (invalid 2D-weights with wrong row number; individual)
+            np.array(
+                [[1.0, 2.0, 3.0], [4.0, 5.0, 6.0], [7.0, 8.0, 9.0], [10.0, 11.0, 12.0]]
+            ),
+            ValueError,
+        ),
+        (
+            # Number 10 (invalid 2D-weights with wrong row and column number;
+            # individual)
+            np.array([[1.0, 2.0, 3.0, 4.0], [5.0, 6.0, 7.0, 8.0]]),
+            ValueError,
+        ),
+        (
+            # Number 10 (invalid 2D-weights with wrong row and column number;
+            # individual)
+            np.array([[1.0, 2.0], [3.0, 4.0]]),
+            ValueError,
+        ),
+        (
+            # Number 11 (invalid 2D-weights with wrong row and column number;
+            # individual)
+            np.array(
+                [
+                    [1.0, 2.0, 3.0, 4.0],
+                    [5.0, 6.0, 7.0, 8.0],
+                    [9.0, 10.0, 11.0, 12.0],
+                    [13.0, 14.0, 15.0, 16.0],
+                ]
+            ),
+            ValueError,
+        ),
+        (
+            # Number 12 (invalid 2D-weights with wrong row and column number;
+            # individual)
+            np.array([[1.0, 2.0], [3.0, 4.0], [5.0, 6.0], [7.0, 8.0]]),
+            ValueError,
+        ),
+        (  # Number 13 (invalid 1D-weights with negative entry; for all)
+            np.array([1.0, 2.0, -1_000.0]),
+            ValueError,
+        ),
[…]
