### 2024-05-13 85563c1
refactor: went from index-based to mask-based approach for performance
