### 2024-05-13 1e11bc7
feat: added true spectrum to the test dataset for the automated Whittaker smoothing
