### 2024-05-19 de17316
refactor: went for optimization scheme that can better be ported to low level languages
