### 2024-05-19 435d6b8
refactor: replaced subtraction by not-equals-comparison
