### 2024-05-19 160d3c3
tests/refactor: split up `utils` in functions and models
