+++ tests/test_for_utils/test_banded_linalg.py
@@ -16,7 +16,7 @@ from chemotools.utils.banded_linalg import (
     lu_solve_banded,
     slogdet_lu_banded,
 )
-from tests.test_for_utils.utils import get_banded_slogdet
+from tests.test_for_utils.utils_funcs import get_banded_slogdet
 
 ### Test Suite ###
 
