+++ tests/test_for_utils/utils_funcs.py
@@ -10,16 +10,18 @@ test utilities are working as expected as well.
 
 ### Imports ###
 
-from dataclasses import dataclass
-from math import isnan
-from typing import Tuple
+from math import exp, isnan
+from typing import Tuple, Union
 
 import numpy as np
 from scipy.linalg import eigvals_banded
-from scipy.sparse import csr_matrix
+from scipy.optimize import brute, minimize_scalar
+from scipy.sparse import csc_matrix, csr_matrix
 from scipy.sparse import diags as sp_diags
+from scipy.sparse import linalg as spla
 
-from chemotools.utils import models
+from chemotools.utils.finite_differences import calc_forward_diff_kernel
+from chemotools.utils.whittaker_base import WhittakerLikeSolver
 
 ### Utility Functions ###
 
@@ -32,7 +34,7 @@ def float_is_bit_equal(value: float, reference: float) -> bool:
     Doctests
     --------
     >>> # Imports
-    >>> from tests.test_for_utils.utils import float_is_bit_equal
+    >>> from tests.test_for_utils.utils_funcs import float_is_bit_equal
 
     >>> # Test 1
     >>> float_is_bit_equal(value=1.0, reference=1.0)
@@ -76,7 +78,9 @@ def conv_upper_cho_banded_storage_to_sparse(
     >>> # Imports
     >>> import numpy as np
     >>> from numpy import nan
-    >>> from tests.test_for_utils.utils import conv_upper_cho_banded_storage_to_sparse
+    >>> from tests.test_for_utils.utils_funcs import (
+    ...     conv_upper_cho_banded_storage_to_sparse,
+    ... )
 
     >>> # Generating a set of test matrices
     >>> # Matrix 1
@@ -216,7 +220,9 @@ def conv_lu_banded_storage_to_sparse(
     >>> # Imports
     >>> import numpy as np
     >>> from numpy import nan
-    >>> from tests.test_for_utils.utils import conv_lu_banded_storage_to_sparse
+    >>> from tests.test_for_utils.utils_funcs import (
+    ...     conv_lu_banded_storage_to_sparse,
+    ... )
 
     >>> # Generating a set of test matrices
     >>> # Matrix 1
@@ -373,7 +379,7 @@ def multiply_vect_with_squ_fw_fin_diff_orig_first(
     --------
     >>> # Imports
     >>> import numpy as np
-    >>> from tests.test_for_utils.utils import (
+    >>> from tests.test_for_utils.utils_funcs import (
     ...     multiply_vect_with_squ_fw_fin_diff_orig_first,
     ... )
 
@@ -513,7 +519,7 @@ def multiply_vect_with_squ_fw_fin_diff_transpose_first(
     --------
     >>> # Imports
     >>> import numpy as np
-    >>> from tests.test_for_utils.utils import (
+    >>> from tests.test_for_utils.utils_funcs import (
     ...     multiply_vect_with_squ_fw_fin_diff_transpose_first,
     ... )
 
@@ -642,7 +648,7 @@ def get_banded_slogdet(ab: np.ndarray) -> Tuple[float, float]:
     --------
     >>> # Imports
     >>> import numpy as np
-    >>> from tests.test_for_utils.utils import (
+    >>> from tests.test_for_utils.utils_funcs import (
     ...     conv_upper_cho_banded_storage_to_sparse,
     ...     get_banded_slogdet,
     ... )
@@ -755,57 +761,486 @@ def get_banded_slogdet(ab: np.ndarray) -> Tuple[float, float]:
     return sign, logabsdet
 
 
-### Dataclasses ###
-
-
-@dataclass
-class ExpectedWhittakerSmoothLambda:
[…]
