### 2024-05-19 48da8b5
test/refactor/feat: implemented auto log lambda utility functions based on sparse matrices to fix the ultra long tests with dense matrices
