+++ tests/test_for_utils/utils_funcs.py
@@ -64,9 +64,7 @@ def float_is_bit_equal(value: float, reference: float) -> bool:
     return value == reference
 
 
-def conv_upper_cho_banded_storage_to_sparse(
-    ab: np.ndarray,
-) -> csr_matrix:
+def conv_upper_cho_banded_storage_to_sparse(ab: np.ndarray) -> csr_matrix:
     """
     Converts a banded matrix stored in the upper banded storage used for LAPACK's banded
     Cholesky decomposition to a sparse ``CSR`` matrix.
@@ -207,7 +205,8 @@ def conv_upper_cho_banded_storage_to_sparse(
 
 
 def conv_lu_banded_storage_to_sparse(
-    ab: np.ndarray, l_and_u: Tuple[int, int]
+    ab: np.ndarray,
+    l_and_u: Tuple[int, int],
 ) -> csr_matrix:
     """
     Converts a banded matrix stored in the banded storage used for LAPACK's banded LU
@@ -761,18 +760,18 @@ def get_banded_slogdet(ab: np.ndarray) -> Tuple[float, float]:
     return sign, logabsdet
 
 
-def get_dense_fw_fin_diff_mat(n_data: int, differences: int) -> csc_matrix:
+def get_sparse_fw_fin_diff_mat(n_data: int, differences: int) -> csc_matrix:
     """
     Creates a dense forward finite difference matrix ``D`` of a given difference order.
 
     Doctests
     --------
     >>> # Imports
-    >>> from tests.test_for_utils.utils_funcs import get_dense_fw_fin_diff_mat
+    >>> from tests.test_for_utils.utils_funcs import get_sparse_fw_fin_diff_mat
 
     >>> # Matrix 1
     >>> n_data, differences = 5, 1
-    >>> get_dense_fw_fin_diff_mat(n_data=n_data, differences=differences).toarray()
+    >>> get_sparse_fw_fin_diff_mat(n_data=n_data, differences=differences).toarray()
     array([[-1.,  1.,  0.,  0.,  0.],
            [ 0., -1.,  1.,  0.,  0.],
            [ 0.,  0., -1.,  1.,  0.],
@@ -780,7 +779,7 @@ def get_dense_fw_fin_diff_mat(n_data: int, differences: int) -> csc_matrix:
 
     >>> # Matrix 2
     >>> n_data, differences = 10, 1
-    >>> get_dense_fw_fin_diff_mat(n_data=n_data, differences=differences).toarray()
+    >>> get_sparse_fw_fin_diff_mat(n_data=n_data, differences=differences).toarray()
     array([[-1.,  1.,  0.,  0.,  0.,  0.,  0.,  0.,  0.,  0.],
            [ 0., -1.,  1.,  0.,  0.,  0.,  0.,  0.,  0.,  0.],
            [ 0.,  0., -1.,  1.,  0.,  0.,  0.,  0.,  0.,  0.],
@@ -793,14 +792,14 @@ def get_dense_fw_fin_diff_mat(n_data: int, differences: int) -> csc_matrix:
 
     >>> # Matrix 3
     >>> n_data, differences = 5, 2
-    >>> get_dense_fw_fin_diff_mat(n_data=n_data, differences=differences).toarray()
+    >>> get_sparse_fw_fin_diff_mat(n_data=n_data, differences=differences).toarray()
     array([[ 1., -2.,  1.,  0.,  0.],
            [ 0.,  1., -2.,  1.,  0.],
            [ 0.,  0.,  1., -2.,  1.]])
 
     >>> # Matrix 4
     >>> n_data, differences = 10, 2
[…]
