### 2024-05-19 b924a7d
test/feat: implemented full tests for whittaker base
