+++ tests/test_for_utils/test_whittaker_base.py
@@ -1,0 +1,558 @@
+"""
+Test suite for the utility functions in the :mod:`chemotools.utils.whittaker_base`
+module.
+
+"""
+
+### Imports ###
+
+from math import log
+from typing import Any, Tuple, Type, Union
+
+import numpy as np
+import pytest
+
+from chemotools.utils import models
+from chemotools.utils.whittaker_base.auto_lambda.shared import get_smooth_wrss
+from chemotools.utils.whittaker_base.initialisation import (
+    get_checked_lambda,
+    get_penalty_log_pseudo_det,
+)
+from chemotools.utils.whittaker_base.main import WhittakerLikeSolver
+from chemotools.utils.whittaker_base.misc import get_weight_generator
+from chemotools.utils.whittaker_base.solvers import solve_normal_equations
+from tests.fixtures import noise_level_whittaker_auto_lambda  # noqa: F401
+from tests.fixtures import spectrum_whittaker_auto_lambda  # noqa: F401
+from tests.test_for_utils.utils_funcs import (
+    find_whittaker_smooth_opt_lambda_log_marginal_likelihood,
+)
+from tests.test_for_utils.utils_models import ExpectedWhittakerSmoothLambda
+
+### Type Aliases ###
+
+_RealNumeric = Union[float, int]
+_WhittakerMethod = Union[str, models.WhittakerSmoothMethods]
+_LambdaSpecs = Union[_RealNumeric, Tuple[_RealNumeric, _RealNumeric, _WhittakerMethod]]
+_LambdaSpecsOrFlawed = Union[_LambdaSpecs, str]
+
+### Constants ###
+
+_NAN: float = float("nan")
+
+### Test Suite ###
+
+
+@pytest.mark.parametrize(
+    "combination",
+    [
+        (  # Number 0 (fixed float)
+            100.0,
+            ExpectedWhittakerSmoothLambda(
+                fixed_lambda=100.0,
+                auto_bounds=(_NAN, _NAN),
+                fit_auto=False,
+                method_used=models.WhittakerSmoothMethods.FIXED,
+                log_auto_bounds=(_NAN, _NAN),
+            ),
+        ),
+        (  # Number 1 (fixed integer)
+            100,
+            ExpectedWhittakerSmoothLambda(
+                fixed_lambda=100.0,
+                auto_bounds=(_NAN, _NAN),
+                fit_auto=False,
+                method_used=models.WhittakerSmoothMethods.FIXED,
+                log_auto_bounds=(_NAN, _NAN),
+            ),
+        ),
+        (  # Number 2 (float search space, log marginal likelihood method enum)
+            (100.0, 10_000.0, models.WhittakerSmoothMethods.LOGML),
+            ExpectedWhittakerSmoothLambda(
+                fixed_lambda=_NAN,
+                auto_bounds=(100.0, 10_000.0),
+                fit_auto=True,
+                method_used=models.WhittakerSmoothMethods.LOGML,
+                log_auto_bounds=(log(100.0), log(10_000.0)),
+            ),
+        ),
+        (  # Number 3 (float search space, log marginal likelihood method string)
+            (100.0, 10_000.0, "logml"),
+            ExpectedWhittakerSmoothLambda(
+                fixed_lambda=_NAN,
+                auto_bounds=(100.0, 10_000.0),
+                fit_auto=True,
+                method_used=models.WhittakerSmoothMethods.LOGML,
+                log_auto_bounds=(log(100.0), log(10_000.0)),
+            ),
+        ),
+        (  # Number 4 (integer search space, log marginal likelihood method enum)
+            (100, 10_000, models.WhittakerSmoothMethods.LOGML),
+            ExpectedWhittakerSmoothLambda(
[…]
