### 2024-05-19 a7f07e7
doc: fixed type in docstring
