### 2024-05-19 fc8f698
test/cov: improved test coverage in Whittaker base by excluding edge case and removing dead branch of `if else`
