### 2024-05-19 dce0117
test: added test for `_datacopied`
