### 2024-05-20 d6e5d02
refactor: added defaults to auto-lambda-specs-model
