### 2024-05-20 b34eca1
refactor: made types consistent for whittaker base and estimator
