### 2024-05-20 2bea38b
feat: added author to `whittaker_base`
