### 2024-05-20 9394025
feat: enriched `__init__` of `smooth` by docstring; improved linting
