### 2024-05-20 c408040
refactor: adapted docstrings of WhittakerSmooth; unified type hints; added more detailed code examples
