+++ chemotools/smooth/_whittaker_smooth.py
@@ -14,15 +14,27 @@ adaption of [2]_ was required to make it numerically stable ([3]_).
 
 """
 
-from typing import Optional, Union
+# Authors:
+# Pau Cabaneros
+# Niklas Zell <nik.zoe@web.de>
+
+### Imports ###
+
+from typing import Literal, Optional, Tuple, Union
 
 from numpy import ndarray
 from sklearn.base import BaseEstimator, OneToOneFeatureMixin, TransformerMixin
 from sklearn.utils.validation import check_is_fitted
 
 from chemotools.utils.check_inputs import check_input, check_weights
-from chemotools.utils.models import WhittakerSmoothLambda
-from chemotools.utils.whittaker_base import WhittakerLikeSolver
+from chemotools.utils.types import RealNumeric
+from chemotools.utils.whittaker_base import (
+    WhittakerLikeSolver,
+    WhittakerSmoothLambda,
+    WhittakerSmoothMethods,
+)
+
+### Main Class ###
 
 
 class WhittakerSmooth(
@@ -37,15 +49,20 @@ class WhittakerSmooth(
 
     Parameters
     ----------
-    lam : float or int, default=1e2
-        The lambda parameter to use for the Whittaker smooth.
+    lam : float or int or (float or int, float or int, {"fixed", "logml"} or WhittakerSmoothMethods) or WhittakerSmoothLambda, default=1e2
+        The lambda parameter, a.k.a. the penalty weight, for the Whittaker smooth. In
+        general, higher values lead to smoother results, but changes take effect in a
+        logarithmic rather than linear manner.
+        It may thus not be zero or negative (``< 1e-25``). Also high values combined
+        with high ``differences`` will lead to numerical instability.
+        Please refer to the Notes section for further details.
 
     differences : int, default=1
         The number of differences to use for the Whittaker smooth. If the aim is to
         obtain a smooth estimate of the ``m``-th order derivative, this should be set to
         at least ``m + 2``.
-        Currently, values >= 6 are highly discouraged and might lead to obscured
-        smoothing.
+        Currently, values ``>= 3`` are highly discouraged due to numerical instability
+        that might obscure the smoothing effect.
 
     Attributes
     ----------
@@ -78,23 +95,113 @@ class WhittakerSmooth(
        (2023)
     .. [3] https://math.stackexchange.com/q/4819039/1261538
 
-    """
+    Notes
+    -----
+    For a more convenient usage of the following, it is recommended to import
+    ``WhittakerSmoothLambda`` and ``WhittakerSmoothMethods`` from ``chemotools.smooth``.
+
+    The specification of ``lam`` controls how the lambda parameter is applied/chosen.
+    It may not be zero or negative (``< 1e-25``), but aside from that, it can be
+    specified in three different ways:
+
+    - a scalar: A fixed lambda is used for all signals, which is a good starting
+        point. However, it is important to notice that even similar signals might
+        require quite different lambdas.
+
+        ```python
+        # fixed lambda of 100
+        smoother = WhittakerSmooth(lam=1e2)
+        ```
+
+        Internally, it is represented by the method ``"fixed"`` or ``WhittakerSmoothMethods.FIXED``,
+        but this never has to be specified explicitly.
+
+    - a tuple of two scalars and a string: The scalars serve as the lower and upper
+        bound for searching a lambda according to the method provided by the
+        string.
+        Each signal will then have its own optimized lambda.
+        Currently available methods for automated selection are:
+
+        - ``logml`` or ``WhittakerSmoothMethods.LOGML``: The lambda is chosen by
+            maximizing the log marginal likelihood similar to the optimization used
+            by the ``sklearn.gaussian_process.GaussianProcessRegressor``.
[…]
