### 2024-05-20 0699a7c
refactor: made `whittaker_base` a private utility
