+++ chemotools/smooth/_whittaker_smooth.py
@@ -28,7 +28,7 @@ from sklearn.utils.validation import check_is_fitted
 
 from chemotools.utils.check_inputs import check_input, check_weights
 from chemotools.utils.types import RealNumeric
-from chemotools.utils.whittaker_base import (
+from chemotools.utils._whittaker_base import (
     WhittakerLikeSolver,
     WhittakerSmoothLambda,
     WhittakerSmoothMethods,
