+++ tests/test_for_utils/utils_funcs.py
@@ -21,7 +21,7 @@ from scipy.sparse import diags as sp_diags
 from scipy.sparse import linalg as spla
 
 from chemotools.utils.finite_differences import calc_forward_diff_kernel
-from chemotools.utils.whittaker_base import WhittakerLikeSolver
+from chemotools.utils._whittaker_base import WhittakerLikeSolver
 
 ### Utility Functions ###
 
