### 2024-05-20 6ef864a
refactor: made `banded_linalg` private for now
