+++ chemotools/utils/_whittaker_base/initialisation.py
@@ -10,7 +10,7 @@ from typing import Any, Literal, Tuple, Type, Union
 
 import numpy as np
 
-from chemotools.utils import banded_linalg as bla
+from chemotools.utils import _banded_linalg as bla
 from chemotools.utils import finite_differences as fdiff
 from chemotools.utils import models
 from chemotools.utils.types import RealNumeric
