### 2024-05-20 907baa4
refactor: made utility `models` and `types` private
