+++ tests/test_for_utils/test_models.py
@@ -10,7 +10,7 @@ from typing import List, Tuple, Type, Union
 
 import pytest
 
-from chemotools.utils import models
+from chemotools.utils import _models
 from tests.test_for_utils.utils_models import ExpectedWhittakerSmoothLambda
 
 ### Type aliases ###
@@ -18,7 +18,7 @@ from tests.test_for_utils.utils_models import ExpectedWhittakerSmoothLambda
 _RealNumeric = Union[float, int]
 _LambdaValueNumeric = Union[_RealNumeric, Tuple[_RealNumeric, _RealNumeric]]
 _LambdaValueNumericOrFlawed = Union[_LambdaValueNumeric, str]
-_WhittakerMethod = Union[str, models.WhittakerSmoothMethods]
+_WhittakerMethod = Union[str, _models.WhittakerSmoothMethods]
 _WhittakerMethodSequence = List[_WhittakerMethod]
 _LambdaTestCombination = Tuple[
     _LambdaValueNumericOrFlawed,
@@ -31,11 +31,11 @@ _LambdaTestCombination = Tuple[
 _NAN: float = float("nan")
 _FIXED_WHITTAKER_METHODS: _WhittakerMethodSequence = [
     "fixed",
-    models.WhittakerSmoothMethods.FIXED,
+    _models.WhittakerSmoothMethods.FIXED,
 ]
 _LOGML_WHITTAKER_METHODS: _WhittakerMethodSequence = [
     "logml",
-    models.WhittakerSmoothMethods.LOGML,
+    _models.WhittakerSmoothMethods.LOGML,
 ]
 # NOTE: "aauto" is not a typo, but helps to not confuse it with "all"
 _aauto_whittaker_methods: _WhittakerMethodSequence = _LOGML_WHITTAKER_METHODS + []
@@ -57,7 +57,7 @@ _all_whittaker_methods: _WhittakerMethodSequence = (
                 fixed_lambda=100.0,
                 auto_bounds=(_NAN, _NAN),
                 fit_auto=False,
-                method_used=models.WhittakerSmoothMethods.FIXED,
+                method_used=_models.WhittakerSmoothMethods.FIXED,
                 log_auto_bounds=(_NAN, _NAN),
             ),
         ),
@@ -68,7 +68,7 @@ _all_whittaker_methods: _WhittakerMethodSequence = (
                 fixed_lambda=100.0,
                 auto_bounds=(_NAN, _NAN),
                 fit_auto=False,
-                method_used=models.WhittakerSmoothMethods.FIXED,
+                method_used=_models.WhittakerSmoothMethods.FIXED,
                 log_auto_bounds=(_NAN, _NAN),
             ),
         ),
@@ -79,7 +79,7 @@ _all_whittaker_methods: _WhittakerMethodSequence = (
                 fixed_lambda=100.0,
                 auto_bounds=(_NAN, _NAN),
                 fit_auto=False,
-                method_used=models.WhittakerSmoothMethods.FIXED,
+                method_used=_models.WhittakerSmoothMethods.FIXED,
                 log_auto_bounds=(_NAN, _NAN),
             ),
         ),
@@ -90,7 +90,7 @@ _all_whittaker_methods: _WhittakerMethodSequence = (
                 fixed_lambda=100.0,
                 auto_bounds=(_NAN, _NAN),
                 fit_auto=False,
-                method_used=models.WhittakerSmoothMethods.FIXED,
+                method_used=_models.WhittakerSmoothMethods.FIXED,
                 log_auto_bounds=(_NAN, _NAN),
             ),
         ),
@@ -101,7 +101,7 @@ _all_whittaker_methods: _WhittakerMethodSequence = (
                 fixed_lambda=100.000001,
[…]
