### 2024-05-20 face185
refactor: made `finite_differences` a private utility module for now
