### 2024-05-20 6885f7c
style: made variable names of Whittaker base more concise with special focus on single character names; unified variable names globally
