+++ chemotools/utils/_whittaker_base/main.py
@@ -176,8 +176,8 @@ class WhittakerLikeSolver:
     def _solve(
         self,
         lam: float,
-        b_weighted: np.ndarray,
-        w: Union[float, np.ndarray],
+        rhs_b_weighted: np.ndarray,
+        weights: Union[float, np.ndarray],
     ) -> tuple[np.ndarray, _models.BandedSolvers, auto._Factorization]:
         """
         Internal wrapper for the solver methods to solve the linear system of equations
@@ -197,8 +197,8 @@ class WhittakerLikeSolver:
             differences=self.differences_,
             l_and_u=self._l_and_u_,
             penalty_mat_banded=self._penalty_mat_banded_,
-            b_weighted=b_weighted,
-            w=w,
+            rhs_b_weighted=rhs_b_weighted,
+            weights=weights,
             pentapy_enabled=self._pentapy_enabled_,
         )
 
@@ -207,8 +207,8 @@ class WhittakerLikeSolver:
     def _marginal_likelihood_objective(
         self,
         log_lam: float,
-        b: np.ndarray,
-        w: np.ndarray,
+        rhs_b: np.ndarray,
+        weights: np.ndarray,
         w_plus_penalty_plus_n_samples_term: float,
     ) -> float:
         """
@@ -226,8 +226,8 @@ class WhittakerLikeSolver:
         # the solution of the linear system of equations is computed
         b_smooth, _, factorization = self._solve(
             lam=lam,
-            b_weighted=b * w,
-            w=w,
+            rhs_b_weighted=rhs_b * weights,
+            weights=weights,
         )
 
         # finally, the log marginal likelihood is computed and returned (negative since
@@ -239,9 +239,9 @@ class WhittakerLikeSolver:
             lam=lam,
             differences=self.differences_,
             diff_kernel_flipped=self._diff_kernel_flipped_,
-            b=b,
-            b_smooth=b_smooth,
-            w=w,
+            rhs_b=rhs_b,
+            rhs_b_smooth=b_smooth,
+            weights=weights,
             w_plus_penalty_plus_n_samples_term=w_plus_penalty_plus_n_samples_term,
         )
 
@@ -249,8 +249,8 @@ class WhittakerLikeSolver:
 
     def _solve_single_b_fixed_lam(
         self,
-        b: np.ndarray,
-        w: Union[float, np.ndarray],
+        rhs_b: np.ndarray,
+        weights: Union[float, np.ndarray],
         lam: Optional[float] = None,
     ) -> tuple[np.ndarray, float]:
         """
@@ -271,12 +271,12 @@ class WhittakerLikeSolver:
         #       the most efficient way around going into this method in the first place;
         #       in the future this might change and thus, this case is kept for now, but
         #       ignored for coverage
-        if isinstance(w, float):  # pragma: no cover
+        if isinstance(weights, float):  # pragma: no cover
             return (
                 self._solve(
                     lam=lam,
-                    b_weighted=b,
-                    w=w,
+                    rhs_b_weighted=rhs_b,
+                    weights=weights,
                 )[0],
                 lam,
             )
@@ -285,16 +285,16 @@ class WhittakerLikeSolver:
         return (
             self._solve(
                 lam=lam,
-                b_weighted=b * w,
-                w=w,
+                rhs_b_weighted=rhs_b * weights,
+                weights=weights,
             )[0],
             lam,
         )
 
     def _solve_single_b_auto_lam_logml(
         self,
-        b: np.ndarray,
-        w: Union[float, np.ndarray],
+        rhs_b: np.ndarray,
+        weights: Union[float, np.ndarray],
     ) -> tuple[np.ndarray, float]:
         """
[…]
