### 2024-05-20 e78823e
test/feat: tested weight generator more thoroughly
