### 2024-05-20 78610f4
feat: added author name; minor adaption in variable names
