### 2024-05-20 4f1eb69
doc: unified docstrings of Whittaker-like classes
