### 2024-05-20 c224915
doc: updated documentation of `WhittakerSmooth`
