### 2024-05-20 395c7c0
test/doc: clarified doctest origin
