### 2024-05-20 dd4cc7a
test/feat/refactor: added preliminary global/local noise estimation function with tests; minor refactoring of finite differences; improved finite difference tests model handling
