+++ tests/test_for_utils/test_finite_differences.py
@@ -6,39 +6,46 @@ module.
 
 ### Imports ###
 
-from typing import List, Tuple
+from typing import List, Optional
 
 import numpy as np
 import pytest
 
 from chemotools.utils._finite_differences import (
     calc_forward_diff_kernel,
+    estimate_noise_stddev,
     gen_squ_fw_fin_diff_mat_cho_banded,
 )
-from tests.fixtures import reference_finite_differences  # noqa: F401
+from tests.fixtures import noise_level_estimation_refs  # noqa: F401
+from tests.fixtures import noise_level_estimation_signal  # noqa: F401
+from tests.fixtures import reference_forward_finite_differences  # noqa: F401
 from tests.test_for_utils.utils_funcs import (
     conv_upper_cho_banded_storage_to_sparse,
     multiply_vect_with_squ_fw_fin_diff_orig_first,
     multiply_vect_with_squ_fw_fin_diff_transpose_first,
+)
+from tests.test_for_utils.utils_models import (
+    NoiseEstimationReference,
+    RefDifferenceKernel,
 )
 
 ### Test Suite ###
 
 
 def test_forward_diff_kernel(
-    reference_finite_differences: List[Tuple[int, int, np.ndarray]]  # noqa: F811
+    reference_forward_finite_differences: List[RefDifferenceKernel],  # noqa: F811
 ) -> None:
     # each kernel is calculated and compared to the reference
-    for differences, _, reference in reference_finite_differences:
-        kernel = calc_forward_diff_kernel(differences=differences)
+    for ref_diff_kernel in reference_forward_finite_differences:
+        kernel = calc_forward_diff_kernel(differences=ref_diff_kernel.differences)
 
-        assert kernel.size == reference.size, (
-            f"Difference order {differences} with accuracy 1 expected kernel size "
-            f"{reference.size} but got {kernel.size}"
+        assert kernel.size == ref_diff_kernel.size, (
+            f"Difference order {ref_diff_kernel.differences} with accuracy 1 expected "
+            f"kernel size {ref_diff_kernel.size} but got {kernel.size}"
         )
-        assert np.allclose(kernel, reference, atol=1e-8), (
-            f"Difference order {differences} with accuracy 1 expected kernel "
-            f"{reference.tolist()} but got {kernel.tolist()}"
+        assert np.allclose(kernel, ref_diff_kernel.kernel, atol=1e-8), (
+            f"Difference order {ref_diff_kernel.differences} with accuracy 1 expected "
+            f"kernel {ref_diff_kernel.kernel.tolist()} but got {kernel.tolist()}"
         )
 
 
@@ -142,3 +149,191 @@ def test_squ_fw_fin_diff_mat_cho_banded_transpose_first(
     # NOTE: the following check has to be fairly strict when it comes to equivalence
     #       since the NumPy and Chemotools are basically doing the same under the hood
     assert np.allclose(result, result_conv, atol=1e-10, rtol=1e-10)
+
+
+@pytest.mark.parametrize(
+    "series, differences, accuracy, window_size, stddev_min",
+    [
+        (  # Number 0 series is too small for difference kernel
+            np.arange(start=0, stop=5),
+            10,
+            2,
+            3,
+            1e-10,
+        ),
+        (  # Number 1 series is too small for difference kernel
+            np.arange(start=0, stop=5),
+            10,
+            2,
+            None,
+            1e-10,
+        ),
+        (  # Number 2 series is too small for window size
+            np.arange(start=0, stop=5),
+            1,
+            2,
+            11,
+            1e-10,
+        ),
+        (  # Number 3 the difference order is 0
+            np.arange(start=0, stop=10),
+            0,
+            2,
+            3,
+            1e-10,
+        ),
[…]
