### 2024-05-20 77b9f6e
test/refactor: included `power` failure into wrong input test of noise estimation; restructured reference finite difference kernels
