+++ tests/test_for_utils/test_finite_differences.py
@@ -152,13 +152,14 @@ def test_squ_fw_fin_diff_mat_cho_banded_transpose_first(
 
 
 @pytest.mark.parametrize(
-    "series, differences, accuracy, window_size, stddev_min",
+    "series, differences, accuracy, window_size, power, stddev_min",
     [
         (  # Number 0 series is too small for difference kernel
             np.arange(start=0, stop=5),
             10,
             2,
             3,
+            1,
             1e-10,
         ),
         (  # Number 1 series is too small for difference kernel
@@ -166,6 +167,7 @@ def test_squ_fw_fin_diff_mat_cho_banded_transpose_first(
             10,
             2,
             None,
+            1,
             1e-10,
         ),
         (  # Number 2 series is too small for window size
@@ -173,6 +175,7 @@ def test_squ_fw_fin_diff_mat_cho_banded_transpose_first(
             1,
             2,
             11,
+            1,
             1e-10,
         ),
         (  # Number 3 the difference order is 0
@@ -180,6 +183,7 @@ def test_squ_fw_fin_diff_mat_cho_banded_transpose_first(
             0,
             2,
             3,
+            1,
             1e-10,
         ),
         (  # Number 4 the difference order is negative
@@ -187,6 +191,7 @@ def test_squ_fw_fin_diff_mat_cho_banded_transpose_first(
             -1,
             2,
             3,
+            1,
             1e-10,
         ),
         (  # Number 5 the accuracy is odd
@@ -194,6 +199,7 @@ def test_squ_fw_fin_diff_mat_cho_banded_transpose_first(
             2,
             3,
             3,
+            1,
             1e-10,
         ),
         (  # Number 6 the accuracy is odd
@@ -201,6 +207,7 @@ def test_squ_fw_fin_diff_mat_cho_banded_transpose_first(
             2,
             5,
             3,
+            1,
             1e-10,
         ),
         (  # Number 7 the accuracy is 1
@@ -208,6 +215,7 @@ def test_squ_fw_fin_diff_mat_cho_banded_transpose_first(
             2,
             1,
             3,
+            1,
             1e-10,
         ),
         (  # Number 8 the accuracy is 0
@@ -215,6 +223,7 @@ def test_squ_fw_fin_diff_mat_cho_banded_transpose_first(
             2,
             0,
             3,
+            1,
             1e-10,
         ),
         (  # Number 9 the accuracy is negative
@@ -222,6 +231,7 @@ def test_squ_fw_fin_diff_mat_cho_banded_transpose_first(
             2,
             -1,
             3,
+            1,
             1e-10,
         ),
         (  # Number 10 the window size is even
@@ -229,6 +239,7 @@ def test_squ_fw_fin_diff_mat_cho_banded_transpose_first(
             1,
             2,
             6,
+            1,
             1e-10,
         ),
         (  # Number 11 the window size is 0
@@ -236,6 +247,7 @@ def test_squ_fw_fin_diff_mat_cho_banded_transpose_first(
             1,
             2,
             0,
+            1,
             1e-10,
         ),
         (  # Number 12 the window size is negative
@@ -243,20 +255,39 @@ def test_squ_fw_fin_diff_mat_cho_banded_transpose_first(
             1,
             2,
             -1,
+            1,
             1e-10,
         ),
-        (  # Number 13 the minimum standard deviation is zero
+        (  # Number 13 the power is -3
+            np.arange(start=0, stop=10),
+            1,
+            2,
+            3,
+            -3,
+            1e-10,
+        ),
+        (  # Number 14 the power is 3
+            np.arange(start=0, stop=10),
+            1,
+            2,
+            3,
+            3,
+            1e-10,
+        ),
+        (  # Number 15 the minimum standard deviation is zero
[…]
