### 2024-05-21 f859bab
test/feat/refactor: added central finite differences test; made finite difference fixture smarter
