### 2024-05-21 7f13f9b
feat: exposed noise level estimation where it can be useful
