+++ chemotools/smooth/__init__.py
@@ -7,15 +7,18 @@ It defines the estimator classes for smoothing data with an Sklearn-like API:
 - :class:`SavitzkyGolayFilter`
 - :class:`WhittakerSmooth`
 
-as well as auxiliary models to allow for convenient usage of the them:
+as well as auxiliary models and functions to allow for convenient usage of the them:
 
 - :class:`WhittakerSmoothMethods` and :class:`WhittakerSmoothLambda` for the
     :class:`WhittakerSmooth` class.
+- :func:`estimate_noise_stddev` to estimate the local/global noise level of a spectrum
+    which can then be used for weighting the data.
 
 """
 
 ### Imports ###
 
+from chemotools.utils._finite_differences import estimate_noise_stddev  # noqa: F401
 from chemotools.utils._models import (   # noqa: F401
     WhittakerSmoothLambda,
     WhittakerSmoothMethods,
