### 2024-05-21 b042e60
refactor/docs: added kwargs for extrapolator; fixed wrong docstrings and comments; improved docstring
