### 2024-05-23 e6e8405
refactor: renamed `window_size` to `window_length`
