### 2024-05-24 c60c8e3
doc: added proper Notes for weight selection strategies of `WhittakerSmooth`
