### 2024-05-24 109be51
Revert "refactor: renamed `window_size` to `window_length`"
This reverts commit e6e8405933212950e973f4ac1706398906ed2cda.
