### 2024-05-24 124cf50
refactor: renamed `window_length` in docstring to `window_size`
