### 2024-05-29 #6802 @editorialbot
[REVIEW]: chemotools: A Python Package that Integrates Chemometrics and scikit-learn
**Submitting author:** @paucablop (<a href="http://orcid.org/0000-0003-2372-5082">Pau Cabaneros López</a>)
**Repository:** https://github.com/paucablop/chemotools
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.5
