### 2024-06-24 5697fd9
style: [44] - renamed all variables in `utils.banded_linalg` and `utils.finite_differences` and the respective tests to be more concise
