+++ tests/test_for_utils/test_finite_differences.py
@@ -12,18 +12,18 @@ import numpy as np
 import pytest
 
 from chemotools.utils._finite_differences import (
-    calc_central_diff_kernel,
-    calc_forward_diff_kernel,
+    central_finite_difference_coefficients,
     estimate_noise_stddev,
-    gen_squ_fw_fin_diff_mat_cho_banded,
+    forward_finite_difference_kernel,
+    squared_forward_difference_matrix_banded,
 )
-from tests.fixtures import noise_level_estimation_refs  # noqa: F401
+from tests.fixtures import noise_level_estimation_references  # noqa: F401
 from tests.fixtures import noise_level_estimation_signal  # noqa: F401
 from tests.fixtures import reference_finite_differences  # noqa: F401
 from tests.test_for_utils.utils_funcs import (
     conv_upper_cho_banded_storage_to_sparse,
-    multiply_vect_with_squ_fw_fin_diff_orig_first,
-    multiply_vect_with_squ_fw_fin_diff_transpose_first,
+    multiply_vect_with_squared_forward_finite_differences_original_first,
+    multiply_vect_with_squared_forward_finite_differences_transpose_first,
 )
 from tests.test_for_utils.utils_models import (
     NoiseEstimationReference,
@@ -44,19 +44,22 @@ def test_forward_diff_kernel(
     """
 
     # each kernel is calculated and compared to the reference
-    for ref_diff_kernel in reference_finite_differences:
-        kernel = calc_forward_diff_kernel(differences=ref_diff_kernel.differences)
+    for kernel_reference in reference_finite_differences:
+        kernel_chemotools = forward_finite_difference_kernel(
+            differences=kernel_reference.differences
+        )
 
         # first, the size of the kernel is checked ...
-        assert kernel.size == ref_diff_kernel.size, (
-            f"Difference order {ref_diff_kernel.differences} with accuracy 1 - "
-            f"Expected kernel size {ref_diff_kernel.size} but got {kernel.size}"
+        assert kernel_chemotools.size == kernel_reference.size, (
+            f"Difference order {kernel_reference.differences} with accuracy 1 - "
+            f"Expected kernel size {kernel_reference.size} but got "
+            f"{kernel_chemotools.size}"
         )
         # ...  followed by the comparison of the kernel itself
-        assert np.allclose(kernel, ref_diff_kernel.kernel, atol=1e-8), (
-            f"Difference order {ref_diff_kernel.differences} with accuracy 1 - "
-            f"Expected kernel {ref_diff_kernel.kernel.tolist()} but got "
-            f"{kernel.tolist()}"
+        assert np.allclose(kernel_chemotools, kernel_reference.kernel, atol=1e-8), (
+            f"Difference order {kernel_reference.differences} with accuracy 1 - "
+            f"Expected kernel {kernel_reference.kernel.tolist()} but got "
+            f"{kernel_chemotools.tolist()}"
         )
 
 
@@ -71,33 +74,34 @@ def test_central_diff_kernel(
     """
 
     # each kernel is calculated and compared to the reference
-    for ref_diff_kernel in reference_finite_differences:
-        kernel = calc_central_diff_kernel(
-            differences=ref_diff_kernel.differences,
-            accuracy=ref_diff_kernel.accuracy,
+    for kernel_reference in reference_finite_differences:
+        kernel_chemotools = central_finite_difference_coefficients(
+            differences=kernel_reference.differences,
+            accuracy=kernel_reference.accuracy,
         )
 
         # first, the size of the kernel is checked ...
-        assert kernel.size == ref_diff_kernel.size, (
-            f"Difference order {ref_diff_kernel.differences} with accuracy "
-            f"{ref_diff_kernel.accuracy} - Expected kernel size {ref_diff_kernel.size} "
-            f"but got {kernel.size}"
[…]
