+++ tests/test_for_utils/utils_funcs.py
@@ -20,7 +20,7 @@ from scipy.sparse import csc_matrix, csr_matrix
 from scipy.sparse import diags as sp_diags
 from scipy.sparse import linalg as spla
 
-from chemotools.utils._finite_differences import calc_forward_diff_kernel
+from chemotools.utils._finite_differences import forward_finite_difference_kernel
 from chemotools.utils._whittaker_base import WhittakerLikeSolver
 
 ### Utility Functions ###
@@ -177,29 +177,34 @@ def conv_upper_cho_banded_storage_to_sparse(ab: np.ndarray) -> csr_matrix:
     """
 
     # the offset vector is initialised
-    n_diags, n_cols = ab.shape
-    n_diags -= 1
-    main_diag_idx = n_diags
-    offsets = np.arange(start=-n_diags, stop=n_diags + 1, step=1, dtype=np.int64)
+    num_diagonals, num_columns = ab.shape
+    num_diagonals -= 1
+    main_diagonal_index = num_diagonals
+    offsets = np.arange(
+        start=-num_diagonals,
+        stop=num_diagonals + 1,
+        step=1,
+        dtype=np.int64,
+    )
 
     # then, the list of diagonals is created
     diagonals = []
     # the subdiagonals are added first ...
-    for offset in range(n_diags, 0, -1):
-        diagonals.append(ab[main_diag_idx - offset, offset:n_cols])
+    for offset in range(num_diagonals, 0, -1):
+        diagonals.append(ab[main_diagonal_index - offset, offset:num_columns])
 
     # ... followed by the main diagonal ...
-    diagonals.append(ab[main_diag_idx, ::])
+    diagonals.append(ab[main_diagonal_index, ::])
 
     # ... and finally the superdiagonals
-    for offset in range(1, n_diags + 1):
-        diagonals.append(ab[main_diag_idx - offset, offset:n_cols])
+    for offset in range(1, num_diagonals + 1):
+        diagonals.append(ab[main_diagonal_index - offset, offset:num_columns])
 
     # the sparse matrix is created
     return sp_diags(  # type: ignore
         diagonals=diagonals,
         offsets=offsets,  # type: ignore
-        shape=(n_cols, n_cols),
+        shape=(num_columns, num_columns),
         format="csr",
     )
 
@@ -326,12 +331,12 @@ def conv_lu_banded_storage_to_sparse(
     """
 
     # the offset vector is initialised
-    n_low_bands, n_upp_bands = l_and_u
-    main_diag_idx = n_upp_bands
-    n_cols = ab.shape[1]
+    num_subdiagonals, num_superdiagonals = l_and_u
+    main_diagonal_index = num_superdiagonals
+    num_columns = ab.shape[1]
     offsets = np.arange(
-        start=-n_low_bands,
-        stop=n_upp_bands + 1,
+        start=-num_subdiagonals,
+        stop=num_superdiagonals + 1,
         step=1,
         dtype=np.int64,
     )
@@ -339,26 +344,26 @@ def conv_lu_banded_storage_to_sparse(
     # then, the list of diagonals is created
     diagonals = []
     # the subdiagonals are added first ...
-    for offset in range(n_low_bands, 0, -1):
-        diagonals.append(ab[main_diag_idx + offset, 0 : n_cols - offset])
+    for offset in range(num_subdiagonals, 0, -1):
+        diagonals.append(ab[main_diagonal_index + offset, 0 : num_columns - offset])
 
     # ... followed by the main diagonal ...
-    diagonals.append(ab[main_diag_idx, ::])
[…]
