### 2024-06-24 fc8572f
style: [44] - renamed all variables and function of the `_whittaker_base` and all related tests to make them more concise - removed `combination` in `pytest.mark.parametrize` in favour of individual variables which are more concise - made error checks explicit for error messages and not only the Exception type to avoid unintended error behaviour
