+++ chemotools/utils/_whittaker_base/main.py
@@ -17,10 +17,21 @@ import numpy as np
 from chemotools._runtime import PENTAPY_AVAILABLE
 from chemotools.utils import _models
 from chemotools.utils._banded_linalg import LAndUBandCounts
-from chemotools.utils._whittaker_base import auto_lambda as auto
-from chemotools.utils._whittaker_base import initialisation as init
-from chemotools.utils._whittaker_base import solvers
+from chemotools.utils._whittaker_base.auto_lambda import (
+    _Factorization,
+    get_log_marginal_likelihood,
+    get_log_marginal_likelihood_constant_term,
+    get_optimized_lambda,
+)
+from chemotools.utils._whittaker_base.initialisation import (
+    _LambdaSpecs,
+    get_checked_lambda,
+    get_flipped_fw_diff_kernel,
+    get_penalty_log_pseudo_determinant,
+    get_squared_forward_finite_difference_matrix_banded,
+)
 from chemotools.utils._whittaker_base.misc import get_weight_generator
+from chemotools.utils._whittaker_base.solvers import solve_normal_equations
 
 ### Class Implementation ###
 
@@ -34,7 +45,7 @@ class WhittakerLikeSolver:
 
     Attributes
     ----------
-    n_data_ : int
+    num_data_ : int
         The number of data points within the series to smooth. It is equivalent to
         ``n_features_in_``, but it was renamed to be allow for definition after the
         initialisation.
@@ -43,24 +54,24 @@ class WhittakerLikeSolver:
         smooth estimate of the ``m``-th order derivative, this should be set to
         at least ``m + 2``.
         For higher orders, the systems to solve tend to get numerically instable,
-        especially when ``n_data_`` grows large and high values for ``lam_`` are used.
+        especially when ``num_data_`` grows large and high values for ``lam_`` are used.
         Values below 1 are not allowed.
-    _lam_inter_ : WhittakerSmoothLambda
+    _lam_internal_ : WhittakerSmoothLambda
         The internal representation of the lambda parameter to use for the smoothing,
         a.k.a. the penalty weight or smoothing parameter.
         It is internally stored as an instance of the dataclass :class:`WhittakerSmoothLambda`.
     _l_and_u_ : (int, int)
         The number of sub- (first) and superdiagonals (second element) of the final
         matrix to solve for smoothing. Both elements will equal ``differences_``.
-    _diff_kernel_flipped_ : ndarray of shape (0, ) or (differences + 1,)
+    _difference_kernel_flipped_ : ndarray of shape (0, ) or (differences + 1,)
         The flipped kernel to use for the forward finite differences. It is only
         required for the automatic fitting of the lambda parameter by maximizing the log
         marginal likelihood, i.e., when ``lam_ == WhittakerSmoothMethods.LOG_MARGINAL_LIKELIHOOD``.
         Flipping is required due to NumPy's definition of convolution.
-    _penalty_mat_banded_ : ndarray of shape (n_data - differences + 1, n_data - differences + 1)
+    _penalty_matrix_banded_ : ndarray of shape (num_data - differences + 1, num_data - differences + 1)
         The squared forward finite differences matrix ``D.T @ D`` stored in the banded
         storage format used for LAPACK's banded LU decomposition.
-    _penalty_mat_log_pseudo_det_ : float
+    _penalty_matrix_log_pseudo_determinant_ : float
         The natural logarithm of the pseudo-determinant of the squared forward finite
         differences matrix ``D.T @ D`` which is used for the automatic fitting of the
         lambda parameter by maximizing the log marginal likelihood, i.e., when
@@ -97,9 +108,9 @@ class WhittakerLikeSolver:
 
     def _setup_for_fit(
         self,
-        n_data: int,
+        num_data: int,
[…]
