### 2024-06-24 45b8224
style: [44] - added error message pattern matching to the tests for models and input check utility functions
