+++ tests/test_for_utils/test_check_inputs.py
@@ -5,7 +5,7 @@ Test suite for the utility models in the :mod:`chemotools.utils.check_inputs` mo
 
 ### Imports ###
 
-from typing import Optional, Tuple, Type, Union
+from typing import Optional, Tuple, Union
 
 import numpy as np
 import pytest
@@ -16,7 +16,7 @@ from chemotools.utils.check_inputs import check_weights
 
 
 @pytest.mark.parametrize(
-    "combination",
+    "weights, expected_result",
     [
         (  # Number 0 (no weights; for all)
             None,
@@ -36,41 +36,41 @@ from chemotools.utils.check_inputs import check_weights
         ),
         (  # Number 4 (invalid 1D-weights with wrong column number; for all)
             np.array([1.0, 2.0, 3.0, 4.0]),
-            ValueError,
+            ValueError("Weights must have 3 columns, but they have"),
         ),
         (  # Number 5 (invalid 1D-weights with wrong column number; for all)
             np.array([1.0, 2.0]),
-            ValueError,
+            ValueError("Weights must have 3 columns, but they have"),
         ),
         (  # Number 6 (invalid 2D-weights with wrong column number; for all)
             np.array([[1.0, 2.0, 3.0, 4.0]]),
-            ValueError,
+            ValueError("Weights must have 3 columns, but they have"),
         ),
         (  # Number 7 (invalid 2D-weights with wrong column number; for all)
             np.array([[1.0, 2.0]]),
-            ValueError,
+            ValueError("Weights must have 3 columns, but they have"),
         ),
         (  # Number 8 (invalid 2D-weights with wrong row number; individual)
             np.array([[1.0, 2.0, 3.0], [4.0, 5.0, 6.0]]),
-            ValueError,
+            ValueError("Weights must have either 1 or 3 rows, but they have"),
         ),
         (  # Number 9 (invalid 2D-weights with wrong row number; individual)
             np.array(
                 [[1.0, 2.0, 3.0], [4.0, 5.0, 6.0], [7.0, 8.0, 9.0], [10.0, 11.0, 12.0]]
             ),
-            ValueError,
+            ValueError("Weights must have either 1 or 3 rows, but they have"),
         ),
         (
             # Number 10 (invalid 2D-weights with wrong row and column number;
             # individual)
             np.array([[1.0, 2.0, 3.0, 4.0], [5.0, 6.0, 7.0, 8.0]]),
-            ValueError,
+            ValueError("Weights must have either 1 or 3 rows, but they have"),
         ),
         (
             # Number 10 (invalid 2D-weights with wrong row and column number;
             # individual)
             np.array([[1.0, 2.0], [3.0, 4.0]]),
-            ValueError,
+            ValueError("Weights must have either 1 or 3 rows, but they have"),
         ),
         (
             # Number 11 (invalid 2D-weights with wrong row and column number;
@@ -83,80 +83,79 @@ from chemotools.utils.check_inputs import check_weights
                     [13.0, 14.0, 15.0, 16.0],
                 ]
             ),
-            ValueError,
+            ValueError("Weights must have either 1 or 3 rows, but they have"),
         ),
         (
             # Number 12 (invalid 2D-weights with wrong row and column number;
             # individual)
             np.array([[1.0, 2.0], [3.0, 4.0], [5.0, 6.0], [7.0, 8.0]]),
-            ValueError,
+            ValueError("Weights must have either 1 or 3 rows, but they have"),
         ),
         (  # Number 13 (invalid 1D-weights with negative entry; for all)
             np.array([1.0, 2.0, -1_000.0]),
-            ValueError,
+            ValueError("Weights may not be negative, but"),
         ),
[…]
