### 2024-06-24 e905250
feat: [44] added a guideline for proper use of `pytest`
