+++ tests/README.md
@@ -1,0 +1,242 @@
+# ✅❌ Good practices for testing with ``pytest`` 🤔
+
+## 1) ✍️ Type hint your tests
+While this might seem tedious because tests are not the actual code after all, this helps catching errors early when writing tests, especially those one did not think of before. A good type checker will help here by underlining type mismatches in red.
+
+## 2) 🪛 Test your test utility functions
+Some tests require utility functions to be written. Since a utility function is also code that can have bugs, it is important to test it as well.<br>
+They can either be tested with a dedicated test
+
+```python
+def utility_function():
+    return 2.0
+
+def test_utility_function():
+    assert utility_function() == 2.0
+```
+
+or via a doctest that will be included in ``chemotools``' test suite by ``pytest``.
+
+```python
+def utility_function():
+    """
+    Doctests
+    --------
+    >>> utility_function()
+    2.0
+    """
+
+    return 2.0
+```
+
+## 3) 🦾🤖 Don't write the same test twice - use parametrization
+If you have a test that is repeated with different inputs, use parametrisation to avoid writing the same test twice.<br>
+This will make your test suite more readable and maintainable. With the ``pytest.mark.parametrize`` decorator, you can run the same test with different inputs. In the following example, the test will run 5 times with the inputs 1, 4, 9, 16, and 25.
+
+```python
+import pytest
+
+@pytest.mark.parametrize("input", [1, 4, 9, 16, 25])
+def test_is_square(input: int) -> None:
+    assert input ** 0.5 == int(input ** 0.5)
+```
+
+In case you want to test multiple input combinations, you can use multiple wrappings of `@pytest.mark.parametrize`. The next test will run 5 x 5 = 25 combinations of inputs.
+
+```python
+import pytest
+
+@pytest.mark.parametrize("input_2", [1, 4, 9, 16, 25])
+@pytest.mark.parametrize("input_1", [1, 4, 9, 16, 25])
+def test_sum_is_positive(
+    input_1: int,
+    input_2: int
+) -> None:
+    assert input_1 + input_2 > 0
+```
+
+If you need multiple wrappings, but some combinations are not valid, you can use `pytest.skip` to skip the test. The following will run 5 x 5 = 25 tests, but will skip the test when both inputs are 1.
+
+```python
+import pytest
+
+@pytest.mark.parametrize("input_2", [1, 4, 9, 16, 25])
+@pytest.mark.parametrize("input_1", [1, 4, 9, 16, 25])
+def test_sum_is_positive(
+    input_1: int,
+    input_2: int
+) -> None:
+    if input_1 == 1 and input_2 == 1:
+        pytest.skip("This test is not valid")
+
+    assert input_1 + input_2 > 0
+```
+
+Finally, in case your test runs on multiple specific combinations of inputs and expected outputs, you can parametrize the full combination in a ``pytest.mark.parametrize`` decorator.
+
+```python
+import pytest
+
+
+@pytest.mark.parametrize(
+    "input_1, input_2, expected",
+    [
+        (1, 2, 3),
+        (2, 3, 5),
+        (3, 4, 7),
+    ],
+)
+def test_sum_is_correct(
+    input_1: int,
+    input_2: int,
+    expected: int,
+) -> None:
+    assert input_1 + input_2 == expected
+```
+
+## 4) 💣❌ Test your error handling thoroughly based on error messages
+If your function raises an error, you should test that it raises the correct error. You can use the ``pytest.raises`` context manager to check that the function raises the expected error.
+
+```python
+import pytest
+
+def divide(a: int, b: int) -> float:
+    return a / b
+
+def test_divide_by_zero_raises_error() -> None:
+    with pytest.raises(ZeroDivisionError):
+        divide(1, 0)
+```
+
[…]
