### 2024-06-24 3826871
feat: [44] added line on why tests for error handling are mandatory
