### 2024-07-08 #6802 @Kastakin
Hello, I am sorry for my late review!

I really appreciated the topic which is really close to my field of study. I think that the documentation for the project is really well written and clear.

The paper is clear in the exposition and succeeds in being a nice introduction to the package. While it does not give a full start to finish example of application I think that it clearly shows its capabilities for who it is targeted for.

I do not see any problem in publishing it in its current form, it is a 👍 from my part!
