### 2024-07-08 #6802 @bmcfee
Thanks @Kastakin and @vinayak2019 for your reviews!

@paucablop please refer to https://github.com/openjournals/joss-reviews/issues/6802#issuecomment-2214454208 and complete the items listed for the author.

I'm now going down the editor post-review checklist.  I have the following comments on the draft:

- Lines 15-16:
    > "have excelled by developing advanced preprocessing methods designed to attenuate instrument and measuring artifacts from the spectra"

    seem grammatically strange.  Should this be "attenuate instrument *noise*" or something to that effect?
- Line 32:
    > model persistence to standardized files such as joblib or pickle

    joblib is not a file format, so it's not clear what this means.  If you're referring to `joblib.dump`, is this not just an interface to pickle?
- Line 55:
    > asymmetrically reweighed penalized least squares

    should this be *reweighted*?
- Several references (line 110, 113, 121) have slightly broken URLs, eg
    >  https://doi.org/https://doi.org/10.1002/bit.27586

    It's not clear if this is problem on the .bib end or in the rendering of the pdf, but it should be checked.
