### 2024-07-17 #6802 @bmcfee
@paucablop :wave: just checking in to see how you're doing with this?
