### 2024-07-28 #6802 @paucablop
Hi @bmcfee and reveiwers :wave:

Thanks a lot for the help and the comments for the paper. I have updated the manuscript with the suggested improvements.

## __Reviewer and Editors Comments__

1. __Comment 1 [line 15-16]:__ I have updated the text to and substitute the term _artifacts_ by _noise_. I agree it is a better term.
2. __Comment 2 [line 32]:__ I have updated the text to: _using standardized libraries such as ```joblib``` or ```pickle```, as you are right, after reading the documentation I can see that ```joblib.dump``` is an interface to ```pickle```.
3. __Comment 3 [line 55]:__ nicely spotted :rocket:
4. __Comment 4: [references]: I have reviewed all the links in the references, and it seems to be right, I have also tested the links from the generated PDF on two computers (Ubuntu and Windows 10) and opening the links in two different browsers (Edge and Firefox) and they worked fine. I am not so sure how to proceed with this :confused:, could you provide some more info about the issue?

## __Additional Author Tasks After Review is Complete__

- [ x] Double check authors and affiliations (including ORCIDs)

- [ x] Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
