### 2024-07-29 #6802 @bmcfee
Thanks @paucablop !

> 4\. could you provide some more info about the issue?

The issue I was seeing was a redundant `https://doi.org` in the URL.  DOI.org seems to be smart enough to parse this out, so it's not a huge deal, but really it should only be there once.  I expect it could be fixed by replacing

https://github.com/paucablop/chemotools/blob/f0e78800a341281003c9da33dd8bc7860afd6c48/paper/paper.bib#L9C1-L9C47

by

```
    doi = {10.1002/bit.27586},
```
to remove the URL prefix, which will be added back in automatically at compile time.  (Same goes for all other bib entries.)

> * This is the version that will be used in the JOSS paper.

Can you post the version number here so we can update the bot?
