### 2024-07-30 1c7a69a
update doi in bibliography
