+++ paper/paper.bib
@@ -6,7 +6,7 @@
     number = {2},
     pages = {579-591},
     keywords = {continuous-discrete extended Kalman filter, fermentation, hybrid model, lignocellulosic ethanol, spectroscopy},
-    doi = {https://doi.org/10.1002/bit.27586},
+    doi = {10.1002/bit.27586},
     url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/bit.27586},
     eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/bit.27586},
     abstract = {Abstract Operating lignocellulosic fermentation processes to produce fuels and chemicals is challenging due to the inherent complexity and variability of the fermentation media. Real-time monitoring is necessary to compensate for these challenges, but the traditional process monitoring methods fail to deliver actionable information that can be used to implement advanced control strategies. In this study, a hybrid-modeling approach is presented to monitor cellulose-to-ethanol (EtOH) fermentations in real-time. The hybrid approach uses a continuous-discrete extended Kalman filter to reconciliate the predictions of a data-driven model and a kinetic model and to estimate the concentration of glucose (Glu), xylose (Xyl), and EtOH. The data-driven model is based on partial least squares (PLS) regression and predicts in real-time the concentration of Glu, Xyl, and EtOH from spectra collected with attenuated total reflectance mid-infrared spectroscopy. The estimations made by the hybrid approach, the data-driven models and the internal model were compared in two validation experiments showing that the hybrid model significantly outperformed the PLS and improved the predictions of the internal model. Furthermore, the hybrid model delivered consistent estimates even when disturbances in the measurements occurred, demonstrating the robustness of the method. The consistency of the proposed hybrid model opens the doors towards the implementation of advanced feedback control schemes.},
@@ -43,7 +43,7 @@ number = {10},
 pages = {1201-1222},
 year = {2009},
 issn = {0165-9936},
-doi = {https://doi.org/10.1016/j.trac.2009.07.007},
+doi = {10.1016/j.trac.2009.07.007},
 url = {https://www.sciencedirect.com/science/article/pii/S0165993609001629},
 author = {Åsmund Rinnan and Frans van den Berg and Søren Balling Engelsen},
 keywords = {Multiplicative Scatter Correction, Near-infrared spectroscopy, Normalization, Norris-Williams derivation, Pre-processing, Savitzky-Golay derivation, Scatter correction, Spectral derivative, Standard Normal Variate, Review},
@@ -58,7 +58,7 @@ volume = {132},
 pages = {116045},
 year = {2020},
 issn = {0165-9936},
-doi = {https://doi.org/10.1016/j.trac.2020.116045},
+doi = {10.1016/j.trac.2020.116045},
 url = {https://www.sciencedirect.com/science/article/pii/S0165993620302740},
 author = {Puneet Mishra and Alessandra Biancolillo and Jean Michel Roger and Federico Marini and Douglas N. Rutledge},
 keywords = {Multivariate calibration, Ensemble learning, Multi-block analysis, Chemometrics, Preprocessing},
