### 2024-07-30 #6802 @paucablop
Hi @bmcfee!

Thanks for the feedback! :rocket:

:heavy_check_mark: I have fixed the bibliography
:heavy_check_mark: The version number is v0.1.5
:heavy_check_mark: The DOI number is: [10.5281/zenodo.13118393](https://doi.org/10.5281/zenodo.13118393)
:heavy_check_mark: I have ensured that the author information, version number, ORCID and License match!

Many thanks for helping with this process :partying_face:
