### 2024-07-30 #6802 @bmcfee
@paucablop thanks for the corrections!  The bot's identified some missing DOIs in your bibliography that should be added.

For the scikit-learn doi, ACM digital library says 10.5555/1953048.2078195 (this [link](https://dl.acm.org/doi/10.5555/1953048.2078195) works), but doi.org won't resolve it. I've reported it upstream to doi.org, but expect that it won't be fixed any time soon.
