### 2024-07-30 #6802 @danielskatz
This is not actually a DOI, though it looks like it. I'm trying to get ACM to change how it lists these so they don't appear to be DOIs, but it will be a long time before anything changes, I'm afraid. JMLR doesn't provide DOIs for its articles.
