### 2024-07-30 #6802 @bmcfee
Thanks for confirming that @danielskatz !  Doubly annoying that they put "doi" in the URL for resolving the DOI-like-string :rofl:

@paucablop we can skip DOI on that one, but the other ref should be updated.
