### 2024-07-30 #6802 @danielskatz
For future knowledge, anything from ACM that starts with 10.5555 is using their placeholder for things that don't have DOIs...
