+++ paper/paper.bib
@@ -32,6 +32,8 @@
   number={1},
   pages={250--257},
   year={2015},
+  doi = {10.1039/c4an01061b},
+  url = {https://pubs.rsc.org/en/content/articlelanding/2015/an/c4an01061b},
   publisher={Royal Society of Chemistry}
 }
 
