### 2024-07-31 #6802 @bmcfee
Final read-through comments, mainly a few more minor typos/grammatical suggestions:

- line 49: "to generale" ⇒ "to generalize"
- line 57: "create a preprocessing pipelines" ⇒ "create a preprocessing pipeline"
- line 66: it looks like the last line of this code example is truncated.  Perhaps you need to line-wrap it?
- line 70: "based on a given criteria" ⇒ "based on a given criterion" (singular) or "based on given criteria" (plural)
- line 84: "and/or" ⇒ "or" (implied by "either" earlier in sentence)
- line 91: "has being utilized" ⇒ "has been utilized"

Otherwise this looks good!
