### 2024-07-31 #6802 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
