### 2024-08-01 #6802 @Kevin-Mattheus-Moerman
- [ ] Please edit the ZENODO listed version tag to include the `v` so it matches your GitHub version tag which is `v0.1.5`.
